% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Alcohol consumption in Australia from Pandeya et al}
\format{
A data frame with 12 rows and 11 columns:

\describe{
\item{sex}{Whether individuals were male or female}
\item{alcohol_g}{Category for the measure in grams of alcohol consumption}
\item{median_intake}{The median intake for the group}
\item{age_18_24, age_25_34, age_35_44, age_45_54, age_55_64, age_65_74, age_75_plus, age_18_plus}{Proportion of adults in each of the \code{alcohol_g} categories (by age group).}
}
}
\usage{
alcohol
}
\description{
Alcohol consumption among Australian adults in grams/day
}
\references{
Pandeya, Nirmala, et al. "Cancers in Australia in 2010
attributable to the consumption of alcohol." Australian and New Zealand
journal of public health 39.5 (2015): 408-413.
}
\seealso{
\link{dementiarisk} for the relative risks and prevalence estimates
}
\keyword{datasets}
