% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Transform an object into a data.frame}
\arguments{
\item{x}{A \code{pif_class} or \code{cases_class} object.}

\item{...}{Additional parameters (ignored)}
}
\value{
A \code{data.frame} containing the fraction or cases (\code{value}), as well
as \code{standard_deviation}, and confidence interval bounds \code{ci_low} and \code{ci_up}.
The \code{label} is included to differentiate among different fractions or cases.
}
\description{
Gets the fraction (\code{pif()}/\code{paf()}) or the cases
(\code{averted_cases()}/\code{attributable_cases()}) and transforms
them into a \code{data.frame}.
}
\examples{
#Transform one pif
my_pif <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1,
    var_beta = 0.2, label = "My pif")
as.data.frame(my_pif)

#Transform more than one pif
my_paf <- paf(p = 0.5, beta = 1.3, var_p = 0.1, var_beta = 0.2,
    label = "My paf")
as.data.frame(my_pif, my_paf)

#Transform averted cases
cases_1 <- averted_cases(16234, my_paf)
as.data.frame(cases_1)

#Transform multiple averted cases
cases_2 <- averted_cases(87980, my_pif)
as.data.frame(cases_1, cases_2)

}
