% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_rr}
\alias{cancer_rr}
\title{Relative risks for cancer from Pandeya et al}
\format{
A data frame with 24 rows and 5 columns:

\describe{
\item{cancer_type}{The type of cancer associated to the relative risk}
\item{dose}{Dose in grams/day for which the relative risk was estimated}
\item{RR}{Relative risk for cancer given the dose of alcohol}
\item{lower_CI, upper_CI}{Lower and upper bounds for the 95\% confidence interval}
}
}
\usage{
cancer_rr
}
\description{
Relative risks for cancer given different levels of alcohol consumption
}
\references{
Pandeya, Nirmala, et al. "Cancers in Australia in 2010
attributable to the consumption of alcohol." Australian and New Zealand
journal of public health 39.5 (2015): 408-413.
}
\seealso{
\link{dementiarisk} for the relative risks and prevalence estimates
}
\keyword{datasets}
