% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{casecalc}
\alias{casecalc}
\alias{averted_cases}
\alias{attributable_cases}
\title{Attributable cases}
\usage{
averted_cases(
  cases,
  pif,
  variance = 0,
  conf_level = 0.95,
  link = "identity",
  link_inv = NULL,
  link_deriv = NULL
)

attributable_cases(
  cases,
  paf,
  variance = 0,
  conf_level = 0.95,
  link = "identity",
  link_inv = NULL,
  link_deriv = NULL
)
}
\arguments{
\item{cases}{The overall number of cases in the population.}

\item{pif}{A potential impact fraction object created by \code{pif}, \code{paf},
\code{pif_total}, \code{pif_ensemble}, \code{paf_total} or \code{paf_ensemble}.}

\item{variance}{The estimated variance for the cases (default = 0).}

\item{conf_level}{Confidence level for the confidence interval (default 0.95).}

\item{link}{Link function such that the case confidence intervals stay
within the expected bounds (either \code{log} or \code{identity}).}

\item{link_inv}{(Optional). If \code{link} is a function then
yhe inverse of \code{link}. For example if \code{link} is \code{logit} this
should be \code{inv_logit}.}

\item{link_deriv}{Derivative of the \code{link} function. The function tries
to build it automatically from \code{link} using \code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}

\item{paf}{A population attributable fraction object created by \code{paf},
\code{paf_total} or \code{paf_ensemble}.}
}
\value{
A \code{cases_class} object with the attributable cases.
}
\description{
Calculates the number of attributable cases or the number of cases
that would be averted under a counterfactual scenario for a
given fraction (either \code{paf} or \code{pif}).
}
\details{
Negative cases are interpreted as cases that would be caused by
the intervention.
}
\section{Formulas}{


The attributable cases are calculated as:
\deqn{
\text{Attributable cases} = \textrm{PAF} \times \textrm{Cases}
}
and the averted cases are respectively:
\deqn{
\text{Averted cases} = \textrm{PIF} \times \textrm{Cases}
}

The variance is estimated using the product-variance formula:
\deqn{
\textrm{Var}[\text{Averted cases}] =
 \sigma^2_{\textrm{Cases}} \cdot \big( \textrm{PIF}\big)^2 +
 \sigma^2_{\textrm{PIF}} \cdot \big( \textrm{Cases} \big)^2 +
 \sigma^2_{\textrm{PIF}} \cdot \sigma^2_{\textrm{Cases}}
}
}

\examples{
frac <- paf(p = 0.499, beta = log(3.6), var_p = 0.002, var_beta = FALSE)
attributable_cases(100, paf = frac)

frac <- pif(p = 0.499, beta = log(3.6), p_cft = 0.1, var_p = 0.002, var_beta = FALSE)
averted_cases(100, pif = frac)


}
\seealso{
\code{\link[=pif]{pif()}}, \code{\link[=paf]{paf()}}
}
