% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{change_link}
\alias{change_link}
\title{Change the link}
\usage{
change_link(x, link = "identity", link_inv = NULL, link_deriv = NULL)
}
\arguments{
\item{x}{A \code{pif_class} object}

\item{link}{Link function such that the \code{pif} confidence intervals
stays within the expected bounds.}

\item{link_inv}{(Optional). If \code{link} is a function then
yhe inverse of \code{link}. For example if \code{link} is \code{logit} this
should be \code{inv_logit}.}

\item{link_deriv}{Derivative of the \code{link} function. The function tries
to build it automatically from \code{link} using \code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}
}
\value{
A \code{pif_class} object with a different \code{link}.
}
\description{
Change the link function for the potential impact fraction
or population attributable fraction to a different link.
}
\examples{
#A potential impact fraction
pif1 <- pif(p = 0.2, p_cft = 0.1, beta = 1.2, var_p = 0.01,
  var_beta = 0.2)
pif1

#Now change the pif to logit to control the negatives
pif1_logit <- change_link(pif1, link = "logit")
pif1_logit

}
