% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_links}
\alias{check_links}
\title{Function that checks the link}
\usage{
check_links(link, link_deriv, link_inv)
}
\arguments{
\item{link}{Link function such that the \code{pif} confidence intervals
stays within the expected bounds.}

\item{link_deriv}{Derivative of the \code{link} function. The function tries
to build it automatically from \code{link} using \code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}

\item{link_inv}{The inverse of \code{link}. For example if \code{link}
is \code{logit} this should be \code{inv_logit}.}
}
\value{
Invisible. Called for its side effects
}
\description{
Function that checks the link
}
\keyword{internal}
