% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_rr_links}
\alias{check_rr_links}
\title{Function that checks the relative risk link}
\usage{
check_rr_links(rr_link, rr_link_deriv)
}
\arguments{
\item{rr_link}{Link function such that the relative risk is given by
\code{rr_link(beta)}.}

\item{rr_link_deriv}{Derivative of the link function for the relative risk.
The function tries to build it automatically from \code{rr_link} using
\code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}
}
\value{
Invisible. Called for its side effects
}
\description{
Function that checks the relative risk link
}
\keyword{internal}
