% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{children}
\alias{children}
\title{Get the children labels from a \code{pif_class}}
\usage{
children(x, ...)
}
\arguments{
\item{x}{A \code{pif_global_ensemble_class}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A character vector with the names of the fractions that make up \code{x}
}
\description{
Gets the labels of the pif elements that make up a \code{pif_global_ensemble_class}.
}
\note{
The result is \code{NULL} if \code{x} is a \code{pif_atomic_class}
}
