% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{coef}
\alias{coef}
\title{Extract coefficients of a pif object}
\arguments{
\item{object}{A \code{pif_class} object.}

\item{...}{Additional parameters to pass to \code{coef} (ignored)}
}
\description{
Gets the potential impact fraction value
}
\examples{
my_pif <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, sigma_p = 0.1, sigma_beta = 0.2)
coef(my_pif)
}
