% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{confint}
\alias{confint}
\title{Extract confidence intervals}
\arguments{
\item{object}{A \code{pif_class} or a \code{cases_class}.}

\item{level}{Level of confidence desired.}

\item{...}{Additional parameters to pass to \code{confint} (ignored)}
}
\value{
A vector containing the lower and upper bounds of the
confidence interval for the object.
}
\description{
Gets the confidence interval for any fraction or cases.
}
\examples{
my_pif <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1, var_beta = 0.2)
#Default 95\% CI
confint(my_pif)

#Custom 90\% ci:
confint(my_pif, level = 0.90)

}
