% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{cov_ensemble_weights}
\alias{cov_ensemble_weights}
\title{Covariance between the j-th weight of a \code{pif_global_ensemble_class} and
another \code{pif_global_ensemble_class}}
\usage{
cov_ensemble_weights(
  pif1,
  pif2,
  var_weights = NULL,
  var_pif_weights = NULL,
  recursive = !is.null(var_pif_weights)
)
}
\arguments{
\item{pif1}{A \code{pif_global_ensemble_class} from which the weight is taken.}

\item{pif2}{A \code{pif_global_ensemble_class} or \code{pif_atomic_class} to compute the covariance}

\item{var_weights}{Covariance vector between the weights of
\code{pif1} and the j-th weight of \code{pif2}.}

\item{j}{Weight indicator (covariance is for \code{weights(pif1)[j]})}

\item{sigma_pif_weights}{Covariance vector between the potential
impact fractions in \code{pif1} and the weights in \code{pif2}.}
}
\description{
Calculates the covariance of the jth-weight
of a potential impact fraction of \code{pif_global_ensemble_class}
with a second \code{pif_global_ensemble_class} or \code{pif_atomic_class}.
}
\note{
The model currently works under the assumption that all \code{pif_atomic_class}
are independent from weights. Hence will return \code{0} if \code{pif2} is an atomic pif.
}
\section{Formula}{

Given a \code{pif_global_ensemble}:
\deqn{
\widehat{\textrm{PIF}}_{1} =
g^{-1}\Bigg(\sum\limits_{i=1}^{M_1} g(\hat{q}_i \cdot \widehat{\textrm{PIF}}_{1,i})\Bigg)
}
and a second \code{pif_global_ensemble}:
\deqn{
\widehat{\textrm{PIF}}_{2} = h^{-1}\Bigg(\sum\limits_{j=1}^{M_2} g(\hat{w}_j \cdot
\widehat{\textrm{PIF}}_{2,j})\Bigg)
}
This function computes the covariance between the first impact fraction and the
jth weight \eqn{w_j} of the second fraction by computing:
\deqn{
\operatorname{Cov}(\widehat{\textrm{PIF}}_A, \hat{w}_j) \approx
\frac{1}{g'\big(\widehat{\textrm{PIF}}_A\big)}\sum\limits_{i=1}^{M_1}g'(\hat{q}_i
\cdot \widehat{\textrm{PIF}}_{A,i}) \Bigg[ \hat{q} \operatorname{Cov}\Big(
\widehat{\textrm{PIF}}_{A,i}, \hat{w}_j\Big) +  \widehat{\textrm{PIF}}_{A,i} \operatorname{Cov}\Big( \hat{q}_j,
\hat{w}_j\Big) \Bigg]
}
where \eqn{\widehat{\textrm{PIF}}_{1,:} = (\widehat{\textrm{PIF}}_{1,1},
\widehat{\textrm{PIF}}_{1,2}, \dots, \widehat{\textrm{PIF}}_{1,M_1})^{\top}}
}

\seealso{
\code{\link[=cov_atomic_pif]{cov_atomic_pif()}}, \code{\link[=cov_ensemble_atomic]{cov_ensemble_atomic()}}
}
\keyword{internal}
