% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{cov_total_pif}
\alias{cov_total_pif}
\title{Covariance function for a \code{pif_global_ensemble_class}}
\usage{
cov_total_pif(
  pif1,
  pif2,
  var_p = NULL,
  var_beta = NULL,
  var_weights = NULL,
  quiet = FALSE
)
}
\arguments{
\item{pif1}{Either a \code{pif_atomic_class} or a \code{pif_global_ensemble_class}}

\item{pif2}{Either a \code{pif_atomic_class} or a \code{pif_global_ensemble_class}}

\item{var_p}{covariance matrix for the prevalences in both \code{pif1} and \code{pif2}.
Default \code{NULL} (no correlation).}

\item{var_beta}{covariance matrix for the parameter \code{beta} in both \code{pif1}
and \code{pif2}. Default \code{NULL} (no correlation).}

\item{var_weights}{Covariance between the weights of all the fractions
contained in \code{pif1} and all the fractions contained in \code{pif2}.}

\item{quiet}{Whether to throw warnings or other messages.}
}
\description{
Recursively obtains the covariance matrix of a \code{pif_global_ensemble_class}
}
\section{Computation}{

This computes:
\deqn{
\operatorname{Cov}(\widehat{\textrm{PIF}}_{A}, \widehat{\textrm{PIF}}_{B})  \approx
\frac{1}{g'\big(\widehat{\textrm{PIF}}_A\big)}\sum\limits_{i=1}^{M_1}g'(\hat{q}_i \cdot
\widehat{\textrm{PIF}}_{A,i}) \Bigg[ \mathbb{E}[\hat{q}_i] \operatorname{Cov}\Big( \textrm{PIF}_{A,i},
\widehat{\textrm{PIF}}_{B}\Big) +  \mathbb{E}[\textrm{PIF}_{A,i}] \operatorname{Cov}\Big( \hat{q}_i,
\widehat{\textrm{PIF}}_{B}\Big) \Bigg]
}
where \eqn{\operatorname{Cov}\Big( \textrm{PIF}_{A,i},
\widehat{\textrm{PIF}}_{B}\Big)} is computed as in \code{\link[=cov_ensemble_atomic]{cov_ensemble_atomic()}}
in the case \eqn{\widehat{\textrm{PIF}}_{A,i}} is atomic or with the same formula
in the case it is an ensemble. The expression \eqn{\operatorname{Cov}\Big( \hat{q}_i,
\widehat{\textrm{PIF}}_{B}\Big) } is computed as in  \code{\link[=cov_ensemble_weight]{cov_ensemble_weight()}}
for each weight \eqn{q_j}.
}

\section{Covariance matrices}{


By default if \code{var_p = NULL} is specified this assumes the parameters \code{p}
of \code{pif1} and \code{pif2} are uncorrelated. However, if \code{pif1} and \code{pif2} share
the same prevalence estimates
(i.e. share the same \code{p}s, then the user should set \code{uncorrelated_p = FALSE}
to account for that correlation).

The same thing happens with \code{var_beta}. If no \code{var_beta} is specified (set to \code{NULL})
then the assumption is that the \code{beta} parameters from \code{pif1} and
from \code{pif2} are uncorrelated (unless \code{uncorrelated_beta} is set to \code{FALSE}).

If the user provides a covariance matrix for \code{var_p} then \code{uncorrelated_p}
is disregarded. Similarly, if the user provides \code{var_beta} then
\code{uncorrelated_beta} is ignored.
}

\seealso{
\code{\link[=from_parameters_covariance_p_component]{from_parameters_covariance_p_component()}}, \code{\link[=cov_atomic_pif]{cov_atomic_pif()}},
\code{\link[=cov_ensemble_atomic]{cov_ensemble_atomic()}}, \code{\link[=cov_ensemble_weight]{cov_ensemble_weight()}}
}
\keyword{internal}
