% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-from_parameters_covariance.R
\name{covariance_from_parameters}
\alias{covariance_from_parameters}
\alias{from_parameters_covariance_p_component}
\alias{from_parameters_covariance_beta_component}
\alias{from_parameters_pif_covariance}
\alias{from_parameters_pif_variance}
\title{Covariance components}
\usage{
from_parameters_covariance_p_component(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  var_p,
  upper_bound
)

from_parameters_covariance_beta_component(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  rr_link_deriv_vals1,
  rr_link_deriv_vals2,
  var_beta,
  upper_bound
)

from_parameters_pif_covariance(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  rr_link_deriv_vals1,
  rr_link_deriv_vals2,
  var_p,
  var_beta,
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)

from_parameters_pif_variance(
  p,
  p_cft,
  rr,
  rr_link_deriv_vals,
  var_p,
  var_beta,
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)
}
\arguments{
\item{p1}{Observed proportion exposed in the first potential impact fraction}

\item{p2}{Observed proportion exposed in the first potential impact fraction}

\item{p1_cft}{Counterfactual proportion exposed in the first potential
impact fraction}

\item{p2_cft}{Counterfactual proportion exposed in the first potential
impact fraction}

\item{rr1}{Relative risk in the first potential impact fraction}

\item{rr2}{Relative risk in the second potential impact fraction}

\item{var_p}{Covariance matrix with the entry \code{var_p[i,j]} corresponding
to the covariance between \code{p1[i]} and \code{p2[j]}.}

\item{upper_bound}{Whether the variance should be calculated or an upper
bound for the variance (assuming perfect correlation).}

\item{rr_link_deriv_vals1}{Values for the derivative of the relative
risk function evaluated at \code{theta} (the relative risk parameter).}

\item{rr_link_deriv_vals2}{Values for the derivative of the relative
risk function evaluated at \code{theta} (the relative risk parameter).}

\item{var_beta}{Estimate of the link_covariance matrix of \code{beta} where the entry
\code{var_beta[i,j]} represents the link_covariance between \code{beta[i]} and \code{beta[j]}.}

\item{upper_bound_p}{Whether the values for the \code{p} component
of the link_variance should be approximated by an upper bound.}

\item{upper_bound_beta}{Whether the values for the \code{beta} component
of the link_variance should be approximated by an upper bound.}

\item{p}{Prevalence (proportion) of the exposed individuals for
each of the \code{N} exposure levels.}

\item{p_cft}{Counterfactual prevalence (proportion) of the exposed
individuals for each of the \code{N} exposure levels.}

\item{rr}{The relative risk for each of the exposure levels.}

\item{rr_link_deriv_vals}{The derivative of the relative risk function \code{g}
with respect to the parameter \code{beta} evaluated at \code{beta}.}
}
\value{
The covariance component for either \code{p} or \code{beta}
}
\description{
Calculates the covariance \code{p} or \code{beta} component between two
potential impact fractions as well as the complete variance
and covariance from parameters.
}
\note{
This estimation does not require a \code{pif_class} object
but instead is built from parameters.
}
\section{Formulas}{

The following represents the covariance components.
For \code{p}:

\deqn{
\text{CC}_{p} = \dfrac{
 \partial \textrm{PIF}_i}{\partial p_i}^{\top}
 \textrm{covariance}\big( \hat{p}_i, \hat{p}_j\big)
 \dfrac{\partial \textrm{PIF}_j}{\partial p_j}
}

and for \code{beta}:

\deqn{
\text{CC}_{\beta} =  \dfrac{
\partial \textrm{PIF}_i}{\partial \beta_i}^{\top}
\textrm{covariance}\big( \hat{\beta}_i, \hat{\beta}_j\big)
\dfrac{\partial \textrm{PIF}_j}{\partial \beta_j}
}

The total covariance is approximated by the sum:
\deqn{
\text{Cov}(\text{PIF}_i, \text{PIF}_j) \approx \text{CC}_{p} + \text{CC}_{\beta}
}
}

\seealso{
\link{derivatives} for the definition of the derivatives involved
in the formulas.
}
\keyword{internal}
