% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-covariance_structures.R
\name{covariance_structures}
\alias{covariance_structures}
\alias{covariance_structure}
\alias{covariance_structure2}
\alias{default_weight_covariance_structure}
\alias{default_weight_covariance_structure2}
\alias{default_parameter_covariance_structure}
\alias{default_parameter_covariance_structure2}
\alias{default_pif_covariance_structure}
\alias{default_pif_covariance_structure2}
\alias{default_weight_pif_covariance_structure}
\alias{default_weight_pif_covariance_structure2}
\title{Default covariance structures}
\usage{
covariance_structure(pif, is_variance = FALSE, warning = TRUE)

covariance_structure2(pif1, pif2, warning = TRUE)

default_weight_covariance_structure(pif, is_variance = FALSE, warning = TRUE)

default_weight_covariance_structure2(pif1, pif2, warning = TRUE)

default_parameter_covariance_structure(
  pif,
  parameter = "p",
  is_variance = FALSE,
  warning = TRUE
)

default_parameter_covariance_structure2(
  pif1,
  pif2,
  parameter = "p",
  warning = TRUE
)

default_pif_covariance_structure(pif, is_variance = FALSE, warning = TRUE)

default_pif_covariance_structure2(pif1, pif2, warning = TRUE)

default_weight_pif_covariance_structure(
  pif,
  is_variance = FALSE,
  warning = TRUE
)

default_weight_pif_covariance_structure2(pif1, pif2, warning = TRUE)
}
\arguments{
\item{pif}{A potential impact fraction}

\item{is_variance}{Whether the covariance structure corresponds to a \code{variance}
(i.e. when \code{pif1} and \code{pif2} are identical)}

\item{warning}{Whether to throw a warning if \code{pif1} or \code{pif2} have common labels}

\item{pif1}{A potential impact fraction to obtain a covariance structure with \code{pif2}.}

\item{pif2}{A potential impact fraction to obtain a covariance structure with \code{pif1},}

\item{parameter}{Either \code{beta} or \code{p}. Indicating which parameter
we are calculating covariance for.}
}
\value{
A nested list of lists with the entry \verb{[[i]][[j]]} representing
the covariance between elements \code{i} and \code{j}.
}
\description{
These are multidimensional arrays of lists where
the entry \code{list[[i]][[j]]} exists represents the covariance
between elements of the i-th potential impact fraction
and the j-th potential impact fraction. This is particularly
useful when handling ensembles and totals.
}
\note{
The \code{covariance_structure}s ending in \code{2} are meant to
obtain the default covariance structure between two fractions \code{pif1} and \code{pif2}
while the ones that don't end in \code{2} are meant to obtain
the covariance structure of a fraction with itself.
}
\examples{
pif_lead_women <- paf(0.27, 2.2, quiet = TRUE, var_p = 0.001, var_beta = 0.015,
                      label = "Women lead")
pif_rad_women  <- paf(0.12, 1.2, quiet = TRUE, var_p = 0.001, var_beta = 0.022,
                      label = "Women radiation")
pif_women      <- pif_ensemble(pif_lead_women, pif_rad_women, label = "Women",
                               weights = c(0.8, 0.72),
                               var_weights = matrix(c(0.3, 0.1, 0.1, 0.4), ncol = 2))

pif_lead_men   <- paf(0.30, 2.2, quiet = TRUE, var_p = 0.001, var_beta = 0.015,
                       label = "Men lead")
pif_rad_men    <- paf(0.10, 1.2, quiet = TRUE, var_p = 0.001, var_beta = 0.022,
                        label = "Men radiation")
pif_men        <- pif_ensemble(pif_lead_men, pif_rad_men, label = "Men",
                        weights = c(0.65, 0.68),
                        var_weights = matrix(c(0.1, -0.2, -0.2, 0.5), ncol = 2))
pif_tot        <- pif_total(pif_men, pif_women,
                        weights = c(0.49, 0.51), label = "Population",
                        var_weights = matrix(c(0.22, 0.4, 0.4, 0.8), ncol = 2))

#This is the default constructor of a covariance. Use it for custom covariances
covariance_structure(pif_lead_women)
covariance_structure2(pif_lead_women, pif_lead_men)
default_weight_covariance_structure2(pif_men, pif_men)
default_weight_covariance_structure(pif_tot)
default_weight_covariance_structure2(pif_men, pif_women)
default_parameter_covariance_structure(pif_tot, parameter = "beta")
default_parameter_covariance_structure2(pif_lead_women, pif_lead_men, parameter = "beta")
}
