% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltapif-package.R
\docType{package}
\name{deltapif-package}
\alias{deltapif}
\alias{deltapif-package}
\title{deltapif: Estimate Potential Impact and Population Attributable Fractions with Aggregated Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Uses the delta-method to estimate the Potential Impact Fraction (PIF) and the Population Attributable Fraction (PAF) from aggregated data. It creates point-estimates, confidence intervals, and estimates of the variance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rodrigozepeda.github.io/deltapif/}
}

}
\author{
\strong{Maintainer}: Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

}
\keyword{internal}
