% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dementiacov}
\alias{dementiacov}
\title{Relative risk covariance from Lee et al}
\format{
A covariance matrix with 12 rows and 12 columns. Each entry
represents the covariance between them.
}
\usage{
dementiacov
}
\description{
Covariances between the relative risks of Lee et al
}
\references{
Lee, Mark, et al. "Variation in population attributable fraction
of dementia associated with potentially modifiable risk factors by race
and ethnicity in the US." JAMA network open 5.7 (2022): e2219672-e2219672.
}
\seealso{
\link{dementiarisk} for the relative risks and prevalence estimates
}
\keyword{datasets}
