% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dementiarisk}
\alias{dementiarisk}
\title{Exposure and Relative Risk data from Lee et Al}
\format{
A data frame with 12 rows and 11 columns:

\describe{
\item{risk_factor}{The risk factor for dementia}
\item{RR}{The relative risk of dementia associated to the risk factor}
\item{lower_CI, upper_CI}{Lower and upper bounds for the 95\% confidence interval}
\item{logrr}{The logarithm of the relative risk \code{log(RR)}}
\item{sdlog}{The variance of the logarithm of the relative risk \code{variance(log(RR))}}
\item{total}{The proportion of individuals exposed in the overall population}
\item{hispanic}{The proportion of hispanic individuals exposed}
\item{asian}{The proportion of non-hispanic asian individuals exposed}
\item{black}{The proportion of non-hispanic black individuals exposed}
\item{white}{The proportion of non-hispanic white individuals exposed}
}
}
\usage{
dementiarisk
}
\description{
Relative risk and exposure data for dementia risk-factors
}
\references{
Lee, Mark, et al. "Variation in population attributable fraction
of dementia associated with potentially modifiable risk factors by race
and ethnicity in the US." JAMA network open 5.7 (2022): e2219672-e2219672.
}
\seealso{
\link{dementiacov} for the covariance between the risk factors
}
\keyword{datasets}
