% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-links.R
\name{deriv_linkfuns}
\alias{deriv_linkfuns}
\alias{deriv_logit}
\alias{deriv_log_complement}
\alias{deriv_hawkins}
\alias{deriv_identity}
\title{Derivatives of link functions}
\usage{
deriv_logit(pif)

deriv_log_complement(pif)

deriv_hawkins(pif)

deriv_identity(pif)
}
\arguments{
\item{pif}{The value of a potential impact fraction or a
population attributable fraction}
}
\description{
A collection of the derivatives of the link functions
for the potential impact fraction.
}
\details{
The functions programmed are as follows

\deqn{
 \text{deriv\_logit}(\text{PIF}) = \dfrac{
   1
  }{
   \text{PIF} \cdot (1 - \text{PIF})
  },
}

\deqn{
 \text{deriv\_log-complement}(\text{PIF}) = \dfrac{
   1
  }{
   \text{PIF} - 1
  },
}

and

\deqn{
 \text{deriv\_Hawkins}(\text{PIF}) = \dfrac{
   1
  }{
   \sqrt{\text{PIF}^2 + 1}
  },
}
}
\seealso{
\link{linkfuns} for the definition of the link functions and \link{inv_linkfuns}
for the inverses of the link functions.
}
