% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-derivatives.R
\name{derivatives}
\alias{derivatives}
\alias{deriv_pif_p}
\alias{deriv_pif_beta}
\title{Partial derivatives of PIF}
\usage{
deriv_pif_p(p, p_cft, rr, mu_obs = NULL, mu_cft = NULL)

deriv_pif_beta(p, p_cft, rr, rr_link_deriv_vals, mu_obs = NULL, mu_cft = NULL)
}
\arguments{
\item{p}{Prevalence (proportion) of the exposed individuals for
each of the \code{N} exposure levels.}

\item{p_cft}{Counterfactual prevalence (proportion) of the exposed
individuals for each of the \code{N} exposure levels.}

\item{rr}{The relative risk for each of the exposure levels.}

\item{mu_obs}{The average value of the relative risk in the observed population.}

\item{mu_cft}{The average value of the counterfactual relative risk in the population.}

\item{rr_link_deriv_vals}{The derivative of the relative risk function \code{g}
with respect to the parameter \code{beta} evaluated at \code{beta}.}
}
\value{
The partial derivative (usually a vector)
}
\description{
Calculates the partial derivatives of a potential impact fraction
with respect to the parameters \code{p} or \code{beta}.
}
\note{
As \code{p} and \code{beta} are usually vectors these are vector-valued
derivatives.
}
\section{Formulas}{

The partial derivative of \code{PIF} with respect to \code{p} is:
\deqn{
\dfrac{\partial \textrm{PIF}}{\partial p} =
\dfrac{\mu^{\text{cft}}}{\big(\mu^{\text{obs}}\big)^2} \cdot \big( \text{RR}(\beta) - 1\big)
}
The partial derivative of \code{PIF} with respect to \code{beta} is:
\deqn{
\dfrac{\partial \textrm{PIF}}{\partial \beta} =
\Bigg(
 \dfrac{
   \mu^{\text{cft}} \cdot p - \mu^{\text{obs}} \cdot p_{*}
 }{
   \Big( \mu^{\text{obs}}\Big)^2
 }
\Bigg)\odot \text{RR}^{\prime}(\beta)
}
with \eqn{\odot} representing the Hadamard (elementwise) product.
}

