% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{flatten}
\alias{flatten}
\title{Function to flatten a \code{pif_global_ensemble_class} and obtain the atomic pifs}
\usage{
flatten(pif)
}
\arguments{
\item{pif}{A potential impact fraction of either \code{pif_atomic_class}
or \code{pif_global_ensemble_class}}
}
\value{
A list of \code{pif_atomic_class} elements
}
\description{
Returns a list of all the potential impact fractions
contained in a \code{pif} object.
}
\examples{
#Potential impact fraction for women
pif_lead_women <- paf(0.27, 2.2, quiet = TRUE, var_p = 0.001)
pif_rad_women  <- paf(0.12, 1.2, quiet = TRUE, var_p = 0.001)
pif_women      <- pif_ensemble(pif_lead_women, pif_rad_women)
pif_lead_men   <- paf(0.30, 2.2, quiet = TRUE, var_p = 0.001)
pif_rad_men    <- paf(0.10, 1.2, quiet = TRUE, var_p = 0.001)
pif_men        <- pif_ensemble(pif_lead_men, pif_rad_men)

flatten(pif_women)

pif_tot <- pif_total(pif_men, pif_women, weights = c(0.49, 0.51))

flatten(pif_tot)

}
