% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{flatten_names}
\alias{flatten_names}
\title{Function to flatten a pif's names}
\usage{
flatten_names(pif, sep = "_>_", add_parents = TRUE)
}
\arguments{
\item{pif}{A potential impact fraction of either \code{pif_atomic_class}
or \code{pif_global_ensemble_class}}
}
\value{
A character vector
}
\description{
Returns a character vector such that the names are
\code{parent_child_grandchild}.
}
\examples{
pif_lead_women <- paf(0.27, 2.2, quiet = TRUE, var_p = 0.001, label = "Women lead")
pif_rad_women  <- paf(0.12, 1.2, quiet = TRUE, var_p = 0.001, label = "Women radiation")
pif_women      <- pif_ensemble(pif_lead_women, pif_rad_women, label = "Women")
pif_lead_men   <- paf(0.30, 2.2, quiet = TRUE, var_p = 0.001, label = "Men lead")
pif_rad_men    <- paf(0.10, 1.2, quiet = TRUE, var_p = 0.001, label = "Men radiation")
pif_men        <- pif_ensemble(pif_lead_men, pif_rad_men, label = "Men")
pif_tot <- pif_total(pif_men, pif_women, weights = c(0.49, 0.51), label = "Population")

flatten_names(pif_lead_women)
flatten_names(pif_tot)

}
