% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{fraction_type}
\alias{fraction_type}
\title{Get the type of the fraction}
\usage{
fraction_type(x)
}
\arguments{
\item{x}{A \code{pif_class} object}
}
\value{
A character either \code{PIF} or \code{PAF} depending on the object
}
\description{
Obtain whether a fraction is a potential impact fraction (PIF) or a
population attributable fraction (PAF)
}
\examples{
#A potential impact fraction
pif1 <- pif(p = 0.2, p_cft = 0.1, beta = 1.2, quiet = TRUE)
fraction_type(pif1)

#A population attributable fraction
paf1 <- paf(p = 0.2, beta = 1.2, quiet = TRUE)
fraction_type(paf1)

}
