% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-getters.R
\name{getters}
\alias{getters}
\alias{get_rr}
\alias{get_mu_obs}
\alias{get_mu_cft}
\alias{get_pif}
\alias{get_link_vals}
\alias{get_link_deriv_vals}
\alias{get_rr_link_deriv_vals}
\alias{get_link_variance}
\alias{get_variance_atomic}
\alias{get_ci}
\alias{get_total_coefs}
\alias{get_ensemble_coefs}
\alias{get_total_pif}
\alias{get_total_type}
\alias{get_ensemble_type}
\alias{get_covariance_total}
\alias{get_ensemble_covariance}
\alias{get_variance_total}
\alias{get_ensemble_variance}
\title{Getters}
\usage{
get_rr(self)

get_mu_obs(self)

get_mu_cft(self)

get_pif(self)

get_link_vals(self)

get_link_deriv_vals(self)

get_rr_link_deriv_vals(self)

get_link_variance(self)

get_variance_atomic(self)

get_ci(self)

get_total_coefs(self)

get_ensemble_coefs(self)

get_total_pif(self)

get_total_type(self)

get_ensemble_type(self)

get_covariance_total(self)

get_ensemble_covariance(self)

get_variance_total(self)

get_ensemble_variance(self)
}
\arguments{
\item{self}{A \code{pif_class} or \code{pif_atomic_class} object created with S7.}
}
\description{
Collection of functions for the potential impact fraction class
so that they work as getters of the properties. Each
function is constructed as \code{get_property}
}
\keyword{internal}
