% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{helper_print_pif_global_ensemble_class}
\alias{helper_print_pif_global_ensemble_class}
\title{Print a \code{pif_global_ensemble_class}}
\usage{
helper_print_pif_global_ensemble_class(x, accuracy)
}
\arguments{
\item{x}{A \code{pif_global_ensemble_class}}

\item{accuracy}{The accuracy parameter for \code{\link[scales:percent_format]{scales::percent}}.}
}
\value{
Called for its side-effects of printing to the console
}
\description{
Prints a \code{pif_global_ensemble_class} object.
}
\keyword{internal}
