% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helpers}
\alias{helpers}
\alias{mu_obs_fun}
\alias{mu_cft_fun}
\alias{pif_fun}
\alias{pif_fun2}
\alias{deriv_pif_p}
\alias{deriv_pif_beta}
\alias{covariance_p_component}
\alias{covariance_beta_component}
\alias{pif_covariance}
\alias{pif_variance}
\alias{pif_ci}
\alias{parse_link}
\alias{parse_inv_link}
\title{Additional helper functions}
\usage{
mu_obs_fun(p, rr)

mu_cft_fun(p_cft, rr)

pif_fun(p, p_cft, rr)

pif_fun2(mu_obs, mu_cft)

deriv_pif_p(p, p_cft, rr, mu_obs = NULL, mu_cft = NULL)

deriv_pif_beta(p, p_cft, rr, rr_link_deriv_vals, mu_obs = NULL, mu_cft = NULL)

covariance_p_component(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  mu_obs1,
  mu_obs2,
  mu_cft1,
  mu_cft2,
  sigma_p,
  upper_bound = FALSE
)

covariance_beta_component(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  rr_link_deriv_vals1,
  rr_link_deriv_vals2,
  mu_obs1,
  mu_obs2,
  mu_cft1,
  mu_cft2,
  sigma_beta,
  upper_bound = FALSE
)

pif_covariance(
  p1,
  p2,
  p1_cft,
  p2_cft,
  rr1,
  rr2,
  rr_link_deriv_vals1,
  rr_link_deriv_vals2,
  mu_obs1,
  mu_obs2,
  mu_cft1,
  mu_cft2,
  sigma_p,
  sigma_beta,
  link_deriv_1,
  link_deriv_2,
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)

pif_variance(
  p,
  p_cft,
  rr,
  rr_link_deriv_vals,
  mu_obs,
  mu_cft,
  sigma_p,
  sigma_beta,
  link_deriv_vals,
  upper_bound_p = FALSE,
  upper_bound_beta = FALSE
)

pif_ci(link_vals, variance, alpha_confint, link_inv)

parse_link(link_name)

parse_inv_link(link_name)
}
\arguments{
\item{p}{Prevalence (proportion) of the exposed individuals for
each of the \code{N} exposure levels.}

\item{rr}{Values for the relative risk at \code{beta} for each of the exposure categories.}

\item{p_cft}{Counterfactual prevalence (proportion) of the exposed
individuals for each of the \code{N} exposure levels.}

\item{mu_obs}{The average relative risk for the observed prevalence}

\item{mu_cft}{The average relative risk for the counterfactual prevalence}

\item{rr_link_deriv_vals}{Values for the relative risk's derivative at \code{beta} for each of the exposure categories.}

\item{sigma_p}{Estimate of the covariance matrix of \code{p} where the entry
\code{sigma_p[i,j]} represents the covariance between \code{p[i]} and \code{p[j]}.}

\item{upper_bound}{Boolean indicating whether we are calculating an
upper bound for the variance (when covariances are unknown) or
if we are calculating the variance (when covariances are known).}

\item{sigma_beta}{Estimate of the covariance matrix of \code{beta} where the entry
\code{sigma_beta[i,j]} represents the covariance between \code{beta[i]} and \code{beta[j]}.}

\item{link_deriv_vals}{The link function's derivative evaluated at the potential impact fraction}

\item{link_vals}{The link function evaluated at the potential impact fraction}

\item{link_inv}{The inverse of \code{link}. For example if \code{link}
is \code{logit} this should be \code{inv_logit}.}

\item{link_name}{The name of the link or a function for the inverse of the link}
}
\description{
Additional helper functions

Compute the mean under the observed prevalence

Compute the potential impact fraction

Compute the potential impact fraction

Partial derivative of PIF respect to p

Partial derivative of PIF respect to beta

Covariance component with respect to p

Covariance component with respect to beta

Covariance between two impact fractions

Variance for a potential impact fraction

Confidence interval for a potential impact fraction

Function to parse the link from a word to a function

Function to parse the inverse link from a word to a function
}
\keyword{Compute}
\keyword{internal}
\keyword{mean}
\keyword{observed}
\keyword{prevalence}
\keyword{the}
\keyword{under}
