% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-links.R
\name{inv_linkfuns}
\alias{inv_linkfuns}
\alias{inv_logit}
\alias{inv_log_complement}
\alias{inv_hawkins}
\title{Inverses of link functions}
\usage{
inv_logit(x)

inv_log_complement(x)

inv_hawkins(x)
}
\arguments{
\item{x}{A value such that \code{inv_link(x)} is a potential impact fraction
or a population attributable fraction.}
}
\description{
A collection of the inverses of the link functions
for the potential impact fraction.
}
\details{
The functions programmed are as follows
\deqn{
 \text{inv\_logit}(\text{PIF}) = \dfrac{
   1
  }{
   1 + \exp(-x)
  },
}
\deqn{
 \text{inv\_log-complement}(\text{PIF}) = 1 - \exp(x),
}
and
\deqn{
 \text{inv\_Hawkins}(\text{PIF}) =
  \frac{1}{2} \exp(-x) \cdot \big(\exp(x) - 1\big) \cdot \big(\exp(x) + 1\big)
}
}
\seealso{
\link{linkfuns} for the definition of the link functions and
\link{deriv_linkfuns} for their derivatives.
}
