% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-links.R
\name{link_parsers}
\alias{link_parsers}
\alias{parse_link}
\alias{parse_inv_link}
\title{Link parsers}
\usage{
parse_link(link_name)

parse_inv_link(link_name)
}
\arguments{
\item{link_name}{The name of the link or a function.}
}
\value{
A function corresponding to the \code{link_name}.
}
\description{
Functions to parse the link (or inverse link) from a word to a function
}
\details{
The following are valid link names:
\describe{
\item{identity}{The function \code{f(x) = x}.with inverse \code{finv(x) = x}}
\item{logit}{The function \code{f(x) = ln(x / (1 - x))} with inverse \code{finv(x) = 1 / (1 + exp(-x))}}
\item{log-complement}{The function \code{f(x) = ln(1 - x)} with inverse \code{finv(x) = 1 - exp(x)}}
\item{hawkins}{The function \code{f(x) = ln(x + sqrt(x^2 + 1))} with inverse \code{finv(x) = 0.5 * exp(-x) * (exp(2 * x) - 1)}}
\item{exponential}{The function \code{f(x) = exp(x)} with inverse \code{f(x) = ln(x)}}
}
}
\note{
If a function is supplied to \code{link_name} the same function is returned
}
\seealso{
\link{linkfuns} and \link{inv_linkfuns}
}
