% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-checkers.R
\name{linkcheck}
\alias{linkcheck}
\alias{check_links}
\alias{check_rr_links}
\title{Function that checks the link}
\usage{
check_links(link, link_deriv, link_inv)

check_rr_links(rr_link, rr_link_deriv)
}
\arguments{
\item{link}{Link function such that the \code{pif} confidence intervals
stays within the expected bounds.}

\item{link_deriv}{Derivative of the \code{link} function. The function tries
to build it automatically from \code{link} using \code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}

\item{link_inv}{The inverse of \code{link}. For example if \code{link}
is \code{logit} this should be \code{inv_logit}.}

\item{rr_link}{Link function such that the relative risk is given by
\code{rr_link(beta)}.}

\item{rr_link_deriv}{Derivative of the link function for the relative risk.
The function tries to build it automatically from \code{rr_link} using
\code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}
}
\value{
Invisible. Called for its side effects
}
\description{
Takes the link or the relative risk link of a \code{pif} or \code{paf} and checks
that it has been correctly specified.
}
\keyword{internal}
