% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{linkfuns}
\alias{linkfuns}
\alias{logit}
\alias{log_complement}
\alias{hawkins}
\title{Link functions}
\usage{
logit(pif)

log_complement(pif)

hawkins(pif)
}
\arguments{
\item{pif}{The value of a potential impact fraction or a
population attributable fraction}
}
\description{
A collection of common link functions, for calculating the variance of
the  potential impact fraction.
}
\details{
The functions programmed are as follows
\deqn{
 \text{logit}(\text{PIF}) = \ln\Bigg(\dfrac{
   \text{PIF}
  }{
   1 - \text{PIF}
  }\Bigg),
}
\deqn{
 \text{log-complement}(\text{PIF}) = \ln\big(1 - \text{PIF}\big),
}
and
\deqn{
 \text{Hawkins}(\text{PIF}) = \ln\Big(\text{PIF} + \sqrt{\text{PIF}^2 + 1}\Big).
}
}
\note{
When used, the variance is calculated for \code{linkfun(pif)}.
}
\seealso{
\link{inv_linkfuns} for their inverses and \link{deriv_linkfuns} for their
derivatives
}
