% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{mu_cft_fun}
\alias{mu_cft_fun}
\title{Compute the mean under the counterfactual prevalence}
\usage{
mu_cft_fun(p_cft, rr)
}
\arguments{
\item{p_cft}{Counterfactual prevalence (proportion) of the exposed
individuals for each of the \code{N} exposure levels.}

\item{rr}{The relative risk for each of the exposure levels.}
}
\value{
The mean relative risk under the counterfactual prevalence
}
\description{
Calculates the mean average relative risk of a population
under the counterfactual prevalence:
\deqn{
E[\text{RR}] = \sum\limits_{i=1}^{n} p_i^{\text{cft}} \cdot \text{RR}_i
}
}
\examples{
\dontrun{
#Consider a popualtion with 3 exposure categories with
#relative risks 1.0, 1.1, 1.3 and prevalences 0.6, 0.3, 0.1
#Notice that the reference relative risk (1) is not added
pval  <- c(0.4, 0.0) #Prevalences on counterfactual scenarios
rrval <- c(1.1, 1.3) #Risks
mu_cft_fun(p_cft = pval, rr = rrval) #Average relative risk
}

}
\seealso{
\code{\link[=mu_obs_fun]{mu_obs_fun()}}
}
\keyword{internal}
