% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{mu_obs_fun}
\alias{mu_obs_fun}
\title{Compute the mean under the observed prevalence}
\usage{
mu_obs_fun(p, rr)
}
\arguments{
\item{p}{Prevalence (proportion) of the exposed individuals for
each of the \code{N} exposure levels.}

\item{rr}{The relative risk for each of the exposure levels.}
}
\value{
The mean relative risk under the observed prevalence
}
\description{
Compute the mean under the observed prevalence
}
\keyword{internal}
