% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{names}
\alias{names}
\title{Get the label of a PIF or PAF}
\arguments{
\item{x}{A \code{pif_class} object.}

\item{...}{Additional parameters (ignored)}
}
\description{
Gets the label of a potential impact fraction or a population
attributable fraction
}
\examples{
#A simple example
my_pif <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1, var_beta = 0.2, label = "Test")
names(my_pif)
names(my_pif) <- "New name"
names(my_pif)

#A pif composed of others
my_pif1 <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1, var_beta = 0.2, label = "Test 1")
my_pif2 <- pif(p = 0.4, p_cft = 0.1, beta = 1.3, var_p = 0.1, var_beta = 0.2, label = "Test 2")
pif_tot <- pif_total(my_pif1, my_pif2, weights = c(0.2, 0.8), label = "Parent")
names(pif_tot)

#Change the name of the parent
names(pif_tot) <- "Total"

#Change the name of all the fractions (3 names)
names(pif_tot) <- c("Total", "pif 1", "pif 2")

#More hierarchy
pif_lead_women <- paf(0.27, 2.2, quiet = TRUE, var_p = 0.001, label = "Women lead")
pif_rad_women  <- paf(0.12, 1.2, quiet = TRUE, var_p = 0.001, label = "Women radiation")
pif_women      <- pif_ensemble(pif_lead_women, pif_rad_women, label = "Women")
pif_lead_men   <- paf(0.30, 2.2, quiet = TRUE, var_p = 0.001, label = "Men lead")
pif_rad_men    <- paf(0.10, 1.2, quiet = TRUE, var_p = 0.001, label = "Men radiation")
pif_men        <- pif_ensemble(pif_lead_men, pif_rad_men, label = "Men")
pif_tot <- pif_total(pif_men, pif_women, weights = c(0.49, 0.51), label = "Population")
names(pif_tot)

#Change the name of all. The order is the same you get when you do `flatten_names`
names(pif_tot)
#flatten_names(pif_tot) #To see in which order to give
names(pif_tot) <- c("All", "Males", "Males Lead", "Males Radiation", "Females", "Females Lead", "Females Radiation")

}
