% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{pif_atomic_ci}
\alias{pif_atomic_ci}
\title{Confidence interval for a potential impact fraction}
\usage{
pif_atomic_ci(link_vals, link_variance, conf_level, link_inv)
}
\arguments{
\item{link_vals}{Values of the link function evaluated at \code{pif} (i.e. \code{link(pif)})}

\item{link_variance}{Link_variance estimate for the linked potential impact
fraction (i.e. for \code{link(pif)})}

\item{conf_level}{Confidence level for the interval}

\item{link_inv}{Inverse of the link function used to compute \code{link_vals} and
\code{link_variance}.}
}
\value{
A vector with the lower and upper bounds of the confidence interval
}
\description{
Confidence interval for a potential impact fraction
}
\seealso{
\link{confint}
}
\keyword{internal}
