% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{pif_class_apply_1st}
\alias{pif_class_apply_1st}
\title{Apply a function to a property of a pif_class or to the first
pif_class available in a pif_total_class}
\usage{
pif_class_apply_1st(x, fun, property)
}
\arguments{
\item{x}{Either a \code{pif_class} or a \code{pif_total_class}}

\item{fun}{A function to apply to the first element of a \code{pif_total_class}}

\item{property}{A property of interest from the \code{pif_class} to extract}
}
\description{
Apply a function to a property of a pif_class or to the first
pif_class available in a pif_total_class
}
\keyword{internal}
