% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pif_list}
\alias{pif_list}
\title{Create a list of \code{pif_list_class} or \code{pif_list} objects:}
\usage{
pif_list(x, ...)
}
\arguments{
\item{x}{A \code{pif_list_class} or \code{pif_list} objects}

\item{...}{Additional \code{pif_list_class} or \code{pif_list} objects}
}
\value{
A \code{pif_list_class} containing all the objects
}
\description{
Create a list of \code{pif_list_class} or \code{pif_list} objects:
}
\keyword{internal}
