% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pif_list_class_apply_1st}
\alias{pif_list_class_apply_1st}
\title{Apply a function to a property of a pif_class or to the first
pif_class available in a pif_list_class}
\usage{
pif_list_class_apply_1st(x, fun, property)
}
\arguments{
\item{Either}{a \code{pif_class} or a \code{pif_list_class}}
}
\description{
Apply a function to a property of a pif_class or to the first
pif_class available in a pif_list_class
}
\keyword{internal}
