% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_total.R
\name{pif_total}
\alias{pif_total}
\title{Total PIF}
\usage{
pif_total(pif1, ..., q, sigma_q)
}
\description{
Total PIF
}
\examples{
beta <- 1.3
sigma_beta <- 0.1
pif1 <- pif(0.5, 0.2, beta, sigma_p = 0.5 * (1 - 0.5) / 100, sigma_beta = sigma_beta)
pif2 <- pif(0.3, 0.1, beta, sigma_p = 0.3 * (1 - 0.3) / 100, sigma_beta = sigma_beta)
pif3 <- pif(0.7, 0.3, beta, sigma_p = 0.7 * (1 - 0.7) / 100, sigma_beta = sigma_beta)
pif_total(pif1, pif2, pif3, q = c(0.3, 0.2, 0.5))
}
