% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-pif_operations.R
\name{pif_validate_ensemble}
\alias{pif_validate_ensemble}
\title{Prepare a global_ensemble}
\usage{
pif_validate_ensemble(
  pif1,
  ...,
  weights,
  sigma_weights,
  sigma_intra_pif_weights,
  conf_level,
  link,
  link_inv,
  link_deriv,
  quiet,
  is_paf = FALSE,
  weights_sum_to_1 = FALSE,
  label
)
}
\arguments{
\item{pif1}{A potential impact fraction (class \code{pif_class})}

\item{...}{The remaining potential impact fractions or (respectively)
population attributable fractions.}

\item{weights}{A vector containing the proportion of the population
for each of the categories (for each of the pifs given).}

\item{sigma_weights}{link_covariance structure for the \code{weights}. Can be \code{0} (default) if
the weights are not random, a vector if only the link_variances of the weights
are available or a link_covariance matrix.}

\item{conf_level}{Confidence level for the confidence interval (default 0.95).}

\item{link}{Link function such that the \code{pif} confidence intervals
stays within the expected bounds.}

\item{link_inv}{The inverse of \code{link}. For example if \code{link}
is \code{logit} this should be \code{inv_logit}.}

\item{link_deriv}{Derivative of the \code{link} function. The function tries
to build it automatically from \code{link} using \code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}

\item{quiet}{Whether to show messages.}

\item{label}{Character identifier for the impact fraction. This is for}
}
\value{
A list of validated values for use in \code{pif_total} or \code{pif_ensemble}.
}
\description{
Helper function that helps validate the inputs on any global ensemble
}
