% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pif.R
\name{pifpaf}
\alias{pifpaf}
\alias{paf}
\alias{pif}
\title{Potential Impact fraction and Population Attributable Fraction}
\usage{
paf(
  p,
  beta,
  var_p = NULL,
  var_beta = NULL,
  rr_link = "exponential",
  rr_link_deriv = NULL,
  link = "log-complement",
  link_inv = NULL,
  link_deriv = NULL,
  conf_level = 0.95,
  quiet = FALSE,
  label = NULL
)

pif(
  p,
  p_cft = rep(0, length(p)),
  beta,
  var_p = NULL,
  var_beta = NULL,
  rr_link = "exponential",
  rr_link_deriv = NULL,
  link = "log-complement",
  link_inv = NULL,
  link_deriv = NULL,
  conf_level = 0.95,
  type = "PIF",
  quiet = FALSE,
  label = NULL
)
}
\arguments{
\item{p}{Prevalence (proportion) of the exposed individuals for
each of the \code{N} exposure levels.}

\item{beta}{Relative risk parameter for which standard deviation is
available (usually its either the relative risk directly or the log
of the relative risk as most RRs, ORs and HRs come from exponential
models).}

\item{var_p}{Estimate of the link_covariance matrix of \code{p} where the entry
\code{var_p[i,j]} represents the link_covariance between \code{p[i]} and \code{p[j]}.}

\item{var_beta}{Estimate of the link_covariance matrix of \code{beta} where the entry
\code{var_beta[i,j]} represents the link_covariance between \code{beta[i]} and \code{beta[j]}.}

\item{rr_link}{Link function such that the relative risk is given by
\code{rr_link(beta)}.}

\item{rr_link_deriv}{Derivative of the link function for the relative risk.
The function tries to build it automatically from \code{rr_link} using
\code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}

\item{link}{Link function such that the \code{pif} confidence intervals
stays within the expected bounds.}

\item{link_inv}{(Optional). If \code{link} is a function then
yhe inverse of \code{link}. For example if \code{link} is \code{logit} this
should be \code{inv_logit}.}

\item{link_deriv}{Derivative of the \code{link} function. The function tries
to build it automatically from \code{link} using \code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}

\item{conf_level}{Confidence level for the confidence interval (default 0.95).}

\item{quiet}{Whether to show messages.}

\item{label}{Character identifier for the impact fraction. This is for}

\item{p_cft}{Counterfactual prevalence (proportion) of the exposed
individuals for each of the \code{N} exposure levels.}

\item{type}{Character either Potential Impact Fraction (\code{PIF}) or
Population Attributable Fraction (\code{PAF})}
}
\description{
Calculates the potential impact fraction \code{pif} or the population
attributable fraction \code{paf} for a categorical exposure considering
an observed prevalence of  \code{p} and a relative risk
(or relative risk parameter) of \code{beta}.
}
\note{
This function assumes \code{p} and \code{beta} have been pre-computed from
the data and the individual-level data are not accessible to the
researchers. If either the data for the individual-level prevalence of
exposure \code{p} or the data for the individual-level risk estimate \code{beta}
can be accessed by the researcher other methods (such as the \code{pifpaf}
package should be preferred).
}
\section{Formulas}{

This function computed the potential impact fraction and its confidence
intervals using Walter's formula:

\deqn{
\textrm{PIF} =
 \dfrac{
   \sum\limits_{i=1}^N p_i \text{RR}_i -
   \sum\limits_{i=1}^N p_i^{\text{cft}} \text{RR}_i
  }{
   \sum\limits_{i=1}^N p_i \text{RR}_i
  }, \quad \text{ and } \quad
\textrm{PAF} =
 \dfrac{
   \sum\limits_{i=1}^N p_i \text{RR}_i - 1
  }{
   \sum\limits_{i=1}^N p_i \text{RR}_i
  }
}

in the case of \code{N} exposure categories which is equivalent to Levine's
formula when there is only \code{1} exposure category:

\deqn{
\textrm{PIF} =
 \dfrac{
   p (\text{RR} - 1) - p^{\text{cft}} (\text{RR} - 1)
  }{
   1 + p (\text{RR} - 1)
  }
  \quad \textrm{ and } \quad
\textrm{PAF} =
 \dfrac{
   p (\text{RR} - 1)
  }{
   1 + p (\text{RR} - 1)
  }
}

The construction of confidence intervals is done via a link function. Its
variance is given by:

\deqn{
\sigma_f^2 = \text{Var}\Big[ f(\textrm{PIF}) \Big] \approx
\Big(f'(\textrm{PIF})\Big)^2 \text{Var}\Big[ \textrm{PIF} \Big]
}
and the intervals are constructed as:
\deqn{
\text{CI} = f^{-1}\Big(  f(\textrm{PIF})  \pm Z_{\alpha/2} \cdot \sigma_f \Big)
}

the function \eqn{f} is called the link function for the PIF, \eqn{f^{-1}}
represents its inverse, and \eqn{f'} its derivative.
}

\section{Link functions for the PIF}{


By default the \code{pif} and \code{paf} calculations use the \code{log-complement} link
which guarantees the impact fractions' intervals have valid values
(between \code{-Inf} and \code{1}).

Depending on the application the following link functions are also
implemented:

\describe{
\item{log-complement}{To achieve fractions between \code{-Inf} and \code{1}. This is
the function \code{f(x) = ln(1 - x)} with inverse \code{finv(x) = 1 - exp(x)}}
\item{logit}{To achieve strictly positive fractions between \code{0} and \code{1}.
This is the function \code{f(x) = ln(x / (1 - x))} with inverse
\code{finv(x) = 1 / (1 + exp(-x))}}
\item{identity}{An approximation for fractions that does not guarantee
confidence intervals in valid regions. This is the function
\code{f(x) = x}.with inverse \code{finv(x) = x}}
\item{hawkins}{Hawkins' fraction for controlling the variance. This is the
function \code{f(x) = ln(x + sqrt(x^2 + 1))} with inverse
\code{finv(x) = 0.5 * exp(-x) * (exp(2 * x) - 1)}}
\item{custom}{User specified (see below).}
}

In general, \code{logit} should be preferred if it is known and certain that the
fractions can only be positive (i.e. when all relative risks (including CIs)
are > 1 and prevalence > 0 and there is an epidemiological / biological
justification).

\strong{Custom link} functions can be implemented as long as they are invertible
in the range of interest by providing:
\describe{
\item{link}{The function}
\item{link_inv}{The inverse function of \code{link}}
\item{link_deriv}{The derivative of \code{link}}
}
If no derivative is provided the package attempts to estimate it
symbolically using \code{\link[Deriv:Deriv]{Deriv::Deriv()}} however there is no guarantee that this
will work for non-standard functions (i.e. not logarithm / trigonometric /
exponential)

As an example considering implementing a square root custom link:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# The link is square root
link      <- sqrt

# Inverse of sqrt(x) is x^2
link_inv  <- function(x) x^2

# The derivative of sqrt(x) is 1/(2 * sqrt(x))
link_deriv <- function(pif) 1 / (2 * sqrt(pif))

# Then the pif can be calculated as:
pif(p = 0.499, beta = 1.6, p_cft = 0.499/2, var_p = 0.1, var_beta = 0.2,
     link = link, link_inv = link_inv, link_deriv = link_deriv)
}\if{html}{\out{</div>}}
}

\section{Link functions for beta}{


By default the \code{pif} and \code{paf} use the \code{exponential} link which means that
the values for \code{beta} are the log-relative risks and the
variance \code{var_beta} corresponds to the log-relative risk's variance.
Depending on the relative risk's source the following options are available:

\describe{
\item{exponential}{For when the relative risks correspond to the
exponential of another parameter, usually called \code{beta}. This is
the exponential function \code{f(beta) = exp(beta)} with inverse
\code{finv(rr) = log(rr) = beta}}
\item{identity}{For when the relative risk and their variance are
reported directly. This is the function \code{f(beta) = beta}
with inverse \code{finv(rr) = rr = beta}}
\item{custom}{User specified (see below).}
}

\strong{Note} that in most cases including contingency tables, Poisson and
Logistic regressions, and Cox Proportional Hazards, the relative risks are
estimated by exponentiating a parameter.

As in the previous section, custom link functions for beta can be implemented
as long as they are invertible in the range of interest by providing the
function \code{rr_link} and its derivative \code{rr_link_deriv}. If no derivative
is provided the package does an attempt to estimate it symbolically
using \code{\link[Deriv:Deriv]{Deriv::Deriv()}} however there is no guarantee that this
will work non-standard functions (i.e. not logarithm / trigonometric /
exponential)
}

\section{Population Attributable Fraction}{


The population attributable fraction corresponds to the potential impact
fraction at the theoretical minimum risk level. It is assumed that the
theoretical minimum risk level is a relative risk of 1. If no
counterfactual prevalence \code{p_cft} is specified, the model computes
the population attributable fraction.
}

\examples{
#EXAMPLE 1: ONE EXPOSURE CATEGORY (CLASSIC LEVIN)
#---------------------------------------------------------------------------
# This example comes from Levin 1953
# Relative risk of lung cancer given smoking was 3.6
# Proportion of individuals smoking where 49.9\%
# Calculates PAF (i.e. counterfactual is no smoking)
paf(p = 0.499, beta = log(3.6))

# Assuming that beta and p had a link_variance
paf(p = 0.499, beta = log(3.6), var_p = 0.001, var_beta = 0.1)

# If the link_variance was to high a logistic transform would be required
# Generates incorrect values for the interval:
paf(p = 0.499, beta = log(3.6), var_p = 0.1, var_beta = 0.3)

# Logit fixes it
paf(p = 0.499, beta = log(3.6), var_p = 0.1, var_beta = 0.3,
    link = "logit", quiet = TRUE)

# If the counterfactual was reducing the smoking population by 1/2
pif(p = 0.499, beta = log(3.6), p_cft = 0.499/2, var_p = 0.001,
    var_beta = 0.1, link = "logit", quiet = TRUE)

#EXAMPLE 2: MULTIPLE EXPOSURE CATEGORIES
#---------------------------------------------------------------------------
#In "Excess Deaths Associated With Underweight, Overweight, and Obesity"
#the PAF of mortality associated to different BMI levels is calculated

#These are the relative risks for age-group 25-59 for each BMI level:
rr <- c("<18.5" = 1.38, "25 to <30" = 0.83 ,
        "30 to <35" = 1.20, ">=35" = 1.83)

#While the prevalences are:
p <- c("<18.5" = 1.9, "25 to <30" = 34.8,
        "30 to <35" = 17.3, ">=35" = 13.3) / 100

#The variance of the relative risk is obtained from the CIs:
var_log_rr <- c("<18.5" = 0.2653156, "25 to <30" =  0.1247604,
                "30 to <35" = 0.1828293, ">=35" = 0.1847374)^2

#Note that we are omitting the group "18.5 to < 25" as it is the
#reference (RR = 1)
paf(p = p, beta = rr, var_beta = var_log_rr, var_p = 0, quiet = TRUE,
    link = "logit")

#We can compute a potential impact fraction of, for example, reducing the
#amount of people over 35 to 0 and having them in the "30 to <35":
p_cft <- c("<18.5" = 1.9, "25 to <30" = 34.8,
            "30 to <35" = 17.3 + 13.3, ">=35" = 0) / 100

#The potential impact fraction is as follows:
pif(p = p, p_cft = p_cft, beta = rr, link = "logit",
  var_beta = var_log_rr, var_p = 0, quiet = TRUE)

}
\seealso{
\code{\link[=pif_total]{pif_total()}}, \code{\link[=pif_ensemble]{pif_ensemble()}}, \code{\link[=paf_total]{paf_total()}}, \code{\link[=paf_ensemble]{paf_ensemble()}},
}
