% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{print_cases_class}
\alias{print_cases_class}
\title{Print a \code{cases_class}}
\usage{
print_cases_class(x, accuracy)
}
\arguments{
\item{x}{A \code{cases_class}}

\item{accuracy}{The accuracy parameter for \code{\link[scales:comma]{scales::comma}}.}
}
\value{
Called for its side-effects of printing to the console
}
\description{
Prints a \code{cases_class} object.
}
\keyword{internal}
