% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{print_pif_class}
\alias{print_pif_class}
\title{Print a \code{pif_class}}
\usage{
print_pif_class(x, accuracy)
}
\arguments{
\item{x}{A \code{pif_class}}

\item{accuracy}{The accuracy parameter for \code{\link[scales:percent_format]{scales::percent}}.}
}
\value{
Called for its side-effects of printing to the console
}
\description{
Prints a \code{pif_class} object.
}
\keyword{internal}
