% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{remove_name_from_covariance}
\alias{remove_name_from_covariance}
\title{Removes the name parent_name from the cov_str list}
\usage{
remove_name_from_covariance(covariance_str, parent_name)
}
\description{
The \code{cov_str} is a list of covariances
}
\keyword{internal}
