% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{rowcol}
\alias{rowcol}
\alias{row_names}
\alias{col_names}
\title{Row and Column names}
\usage{
row_names(x)

col_names(x)
}
\arguments{
\item{x}{A \code{covariance_structure}}
}
\value{
The names of the rows or columns of a \code{covariance_structure}
}
\description{
Retrieve the row and column names of a \code{covariance_structure}
}
\examples{
#' #A pif composed of others
my_pif1 <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1,
                var_beta = 0.2, label = "pif 1")
my_pif2 <- pif(p = 0.4, p_cft = 0.1, beta = 1.3, var_p = 0.1,
                var_beta = 0.2, label = "pif 2")
covst <- covariance_structure2(my_pif1, my_pif2)
row_names(covst)
col_names(covst)


}
