% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-setters.R
\name{setters}
\alias{setters}
\title{Setters}
\arguments{
\item{self}{A \code{pif_class} object created with S7.}
}
\description{
Collection of functions for the potential impact fraction class
so that they work as setters of the properties. Each
function is constructed as \code{set_property}
}
\keyword{internal}
