% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{subset}
\alias{subset}
\title{Subset a \code{covariance_structure}}
\arguments{
\item{x}{A \code{covariance_structure}}

\item{select}{A vector of covariate names to keep in the \code{covariance_structure}.}

\item{cols}{A vector of covariate column names to keep in the \code{covariance_structure}.}

\item{rows}{A vector of covariate row names to keep in the \code{covariance_structure}.}

\item{negate}{If \code{TRUE} subsets the variables that have not been specified}

\item{...}{Additional parameters (ignored)}
}
\description{
Obtains a smaller \code{covariance_structure} with the entries
given by the \code{select} option as a vector.
}
\examples{
pif_lead_women <- paf(0.27, 2.2, quiet = TRUE, var_p = 0.001, var_beta = 0.015,
                      label = "Women lead")
pif_rad_women  <- paf(0.12, 1.2, quiet = TRUE, var_p = 0.001, var_beta = 0.022,
                      label = "Women radiation")
covstr <- default_parameter_covariance_structure2(pif_lead_women, pif_rad_women, parameter = "beta")
subset(covstr, "Women lead")
subset(covstr, "Women lead", negate = TRUE)
subset(covstr, c("Women radiation", "Women lead"))
subset(covstr, 2)
subset(covstr, 1:2)

}
