% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{subset_row}
\alias{subset_row}
\title{Function to get the rows out of a \code{covariance_structure}}
\usage{
subset_row(x, select, ...)
}
\description{
Function to get the rows out of a \code{covariance_structure}
}
\keyword{internal}
