% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-validators.R
\name{validators}
\alias{validators}
\alias{validate_global_ensemble}
\title{Validators
Functions used in the validations of classes}
\usage{
validate_global_ensemble(self)
}
\arguments{
\item{self}{A \code{pif_class} or \code{pif_atomic_class} object created with S7.}
}
\description{
Validators
Functions used in the validations of classes

Validate any generic ensemble including pif total
}
\keyword{internal}
