% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-pif_operations.R
\name{verify_var_pif_weights}
\alias{verify_var_pif_weights}
\title{Verify the variance between pif weights}
\usage{
verify_var_pif_weights(sigma_list, pif_list)
}
\description{
Verify the variance between pif weights
}
