% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{weights}
\alias{weights}
\title{Extract weights of a pif_global_ensemble}
\arguments{
\item{object}{A \code{pif_global_ensemble} object.}

\item{...}{Additional parameters to pass to \code{weights} (ignored)}
}
\description{
Gets the weights of a \code{pif_global_ensemble} object
}
\examples{
my_pif1 <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1, var_beta = 0.2)
my_pif2 <- pif(p = 0.3, p_cft = 0.1, beta = 1.5, var_p = 0.1, var_beta = 0.2)
my_pif  <- pif_total(my_pif1, my_pif2, weights = c(0.8, 0.2))
weights(my_pif)

}
