% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Transform a pif object into a data.frame}
\arguments{
\item{x}{A \code{pif_class} object.}

\item{...}{Additional parameters (ignored)}
}
\description{
Gets the potential impact fraction value, the link_variance and the confidence
interval values
}
\examples{
my_pif <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1, var_beta = 0.2)
as.data.frame(my_pif)
}
