% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{as.list}
\alias{as.list}
\title{Convert a \code{covariance_structure} to \code{list}}
\arguments{
\item{x}{A \code{covariance_structure}}

\item{...}{Additional parameters (ignored)}
}
\description{
Transforms a \code{covariance_structure} into a \code{list}.
}
