% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{as.matrix}
\alias{as.matrix}
\title{Convert a \code{covariance_structure} to \code{matrix}}
\arguments{
\item{x}{A \code{covariance_structure}}
}
\description{
Transforms a \code{covariance_structure} into a \code{matrix}.
}
