% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{as_covstr}
\alias{as_covstr}
\alias{as_covariance_structure}
\title{Transform into a covariance structure}
\usage{
as_covariance_structure(x, col_names = NULL, row_names = NULL, ...)
}
\arguments{
\item{x}{A matrix or a vector}

\item{col_names}{Names to assign to the columns}

\item{row_names}{Names to assign to the rows}

\item{...}{Additional arguments (currently ignored)}
}
\description{
Transforms either a matrix or a vector into a covariance structure.
}
\examples{
mat <- matrix(c(1,3,2,4), ncol = 2,
          dimnames = list(list("pif1", "pif2"), list("pif1", "pif2")))
as_covariance_structure(mat)
}
