% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{children}
\alias{children}
\title{Get the children labels from a \code{pif_class}}
\usage{
children(x, ...)
}
\arguments{
\item{x}{A \code{pif_global_ensemble_class}}
}
\description{
Gets the labels of the elemebts in the \code{pif_list} of an ensemble class.
}
