% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{confint}
\alias{confint}
\title{Extract confidence intervals of a pif object}
\arguments{
\item{object}{A \code{pif_class} object.}

\item{level}{Level of confidence desired.}

\item{...}{Additional parameters to pass to \code{confint} (ignored)}
}
\description{
Gets the confidence interval for the potential impact fraction
}
\examples{
my_pif <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1, var_beta = 0.2)
#Default 95\% CI
confint(my_pif)

#Custom 90\% ci:
confint(my_pif, level = 0.90)
}
