% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{cov_atomic_pif}
\alias{cov_atomic_pif}
\title{Covariance between two atomic pifs}
\usage{
cov_atomic_pif(pif1, pif2, var_p = NULL, var_beta = NULL)
}
\arguments{
\item{pif1}{A \code{pif_atomic_class} object.}

\item{pif2}{A second \code{pif_atomic_class} object.}

\item{var_p}{covariance matrix for the prevalences in both \code{pif1} and \code{pif2}.
Default \code{NULL} (no correlation).}

\item{var_beta}{covariance matrix for the parameter \code{beta} in both \code{pif1}
and \code{pif2}. Default \code{NULL} (no correlation).}
}
\description{
Calculates the approximate covariance between two atomic
pif class expressions using the delta method.
}
\section{Computation}{


The  \code{cov_atomic_pif} computes the approximate covariance between
two potential impact fractions \eqn{\text{PIF}_1} and \eqn{\text{PIF}_2}
as follows:
\deqn{
\text{Cov}\Big(\text{PIF}_1, \text{PIF}_2 \Big) \approx \Bigg(
 \dfrac{\partial \text{PIF}_1}{\partial p_1}^{\top}
 \Sigma_p
 \dfrac{\partial \text{PIF}_2}{\partial p_2} +
 \dfrac{\partial \text{PIF}_1}{\partial \beta_1}^{\top}
 \Sigma_{\beta}
 \dfrac{\partial \text{PIF}_2}{\partial \beta_2}
\Bigg)
}
where \eqn{\Sigma_p} is \code{var_p} and  \eqn{\Sigma_{\beta}} is \code{var_beta}.
The parameters \eqn{p_1} and \eqn{\beta_1} refer to the prevalence (\code{p}) and
relative risk exposure parameter \code{beta} of the first potential
impact fraction \code{pif1} and  \eqn{p_2} and \eqn{\beta_2} to the respective
parameters of \code{pif2}.
}

\examples{
\dontrun{
p1 <- pif(p = 0.5, p_cft = 0.1, beta = 0.2, 0.003, 0.04, label = "Population 1")
p2 <- pif(p = 0.4, p_cft = 0.1, beta = 0.2, 0.003, 0.04, label = "Population 2")

#Positive covariance as program automatically detects same beta and p
cov_atomic_pif(p1, p2)

#Zero covariance as they have no covariates in common
p3 <- pif(p = 0.6, p_cft = 0.1, beta = 0.12, 0.003, 0.04, label = "Population 3")
cov_atomic_pif(p1, p3)

#Covariance has to be given to specify the covariance between p and/or beta parameters
cov_atomic_pif(p1, p3, var_p = 0.12, var_beta = 0.11)

#Works with parameters of different dimensions with entry i-j being the
#covariance between entry p[i] of pif1 and entry p[j] of pif2
#(same with the betas)
p_dim_3 <- paf(p = c(0.5, 0.2, 0.1), beta = c(0.2, 0.1, 0.4),
  label = "Population 1", quiet = TRUE)
p_dim_2 <- pif(p = c(0.4, 0.1), p_cft = c(0.1, 0.05), beta = c(0.3, 0.8),
  label = "Population 2", quiet = TRUE)
cov_atomic_pif(p_dim_2, p_dim_3,
  var_p = matrix(c(0.1, 0.2, 0.4, 0.5, 0.6, 0.08), ncol = 3))
}

}
\seealso{
\code{\link[=from_parameters_covariance_p_component]{from_parameters_covariance_p_component()}} to calculate the covariance
from first principles, \link{cov_ensemble_atomic} and \code{\link[=cov_ensemble_weights]{cov_ensemble_weights()}}
for the covariance between an ensemble fraction and an atomic fraction
and the covariance between the weights of an ensemble fraction
and another fraction.
}
\keyword{internal}
