% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{cov_atomic_pif}
\alias{cov_atomic_pif}
\title{Covariance between two atomic pifs}
\usage{
cov_atomic_pif(pif1, pif2, var_p, var_beta)
}
\arguments{
\item{pif1}{A \code{pif_atomic_class} object.}

\item{pif2}{A second \code{pif_atomic_class} object.}

\item{var_p}{covariance matrix for the prevalences in both \code{pif1} and \code{pif2}.
Default \code{NULL} (no correlation).}

\item{var_beta}{covariance matrix for the parameter \code{beta} in both \code{pif1}
and \code{pif2}. Default \code{NULL} (no correlation).}
}
\description{
Calculates the approximate covariance between two atomic
pif class expressions using the delta method.
}
\section{Computation}{


The  \code{cov_atomic_pif} computes the approximate covariance between
two potential impact fractions \eqn{\text{PIF}_1} and \eqn{\text{PIF}_2}
as follows:
\deqn{
\text{Cov}\Big(\text{PIF}_1, \text{PIF}_2 \Big) \approx \Bigg(
 \dfrac{\partial \text{PIF}_1}{\partial p_1}^{\top}
 \Sigma_p
 \dfrac{\partial \text{PIF}_2}{\partial p_2} +
 \dfrac{\partial \text{PIF}_1}{\partial \beta_1}^{\top}
 \Sigma_{\beta}
 \dfrac{\partial \text{PIF}_2}{\partial \beta_2}
\Bigg)
}
where \eqn{\Sigma_p} is \code{var_p} and  \eqn{\Sigma_{\beta}} is \code{var_beta}.
The parameters \eqn{p_1} and \eqn{\beta_1} refer to the prevalence (\code{p}) and
relative risk exposure parameter \code{beta} of the first potential
impact fraction \code{pif1} and  \eqn{p_2} and \eqn{\beta_2} to the respective
parameters of \code{pif2}.
}

\seealso{
\code{\link[=from_parameters_covariance_p_component]{from_parameters_covariance_p_component()}}
}
\keyword{internal}
