% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{cov_atomic_pif}
\alias{cov_atomic_pif}
\title{Estimate the colink_variance between two pif_atomic objects}
\usage{
cov_atomic_pif(
  pif1,
  pif2,
  var_p = NULL,
  var_beta = NULL,
  independent_p = "guess",
  independent_beta = "guess",
  quiet = FALSE
)
}
\arguments{
\item{pif1}{A \code{pif_atomic_class} object}

\item{pif2}{A \code{pif_atomic_class} object}

\item{var_p}{Colink_variance matrix for the prevalences in all \code{pif1} and
the ones included in \code{...}.}

\item{var_beta}{Colink_variance matrix for the parameter \code{beta} in all \code{pif1}
and the ones included in \code{...}.}

\item{independent_p}{If all the pifs share the same prevalence data. Either
\code{TRUE}, \code{FALSE}, \code{guess} (default) or a matrix. If a matrix is given then
\code{independent_p[i,j] = 1} if the i-th and j-th pifs share the same prevalence data
\code{independent_p[i,j] = 0} if the i-th and j-th pifs don't share the same prevalence data.}

\item{independent_beta}{If all the pifs share the same \code{beta} parameter. Either
\code{TRUE}, \code{FALSE} or \code{guess} (default) or a matrix. If a matrix is given then
\code{independent_beta[i,j] = 1} if the i-th and j-th pifs share the same relative risk parameters
\code{independent_beta[i,j] = 0} if the i-th and j-th pifs don't share the same relative risk parameters.}

\item{quiet}{Whether to throw warnings and other messages}
}
\description{
Estimate the colink_variance between two pif_atomic objects
}
\keyword{internal}
