% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{cov_ensemble_atomic}
\alias{cov_ensemble_atomic}
\title{Covariance between \code{pif_global_ensemble_class} and a \code{pif_atomic}}
\usage{
cov_ensemble_atomic(
  pif_ensemble,
  pif_atomic,
  var_p,
  var_beta,
  sigma_pifs = NULL,
  sigma_weights_pif = NULL,
  recursive = !is.null(sigma_pifs)
)
}
\arguments{
\item{pif_ensemble}{A \code{pif_global_ensemble_class}}

\item{pif_atomic}{A \code{pif_atomic_class}}

\item{sigma_pifs}{Covariance vector between the potential
impact fractions in \code{pif_ensemble} and \code{pif_atomic}.}

\item{sigma_weights_pif}{Covariance vector between the weights in  \code{pif_ensemble} and
the \code{pif_atomic}. For most applications this should be left as \code{NULL} which assumes
independence between the \code{pif_atomic} and the weights used in the \code{pif_ensemble}.}

\item{uncorrelated_pifs}{Whether to "guess" if there is a correlation
between the atomic impact fraction and the impact fraction ensemble.}

\item{uncorrelated_pif_weights}{Whether to "guess" if there is a correlation
between the atomic impact fraction and the weights. Currently ignored.}
}
\description{
Calculates the covariance of a potential impact fraction of a
\code{pif_global_ensemble_class}  with a \code{pif_atomic}
}
\section{Formula}{

Given a \code{pif_global_ensemble}:
\deqn{
\widehat{\textrm{PIF}}_{1} =
g^{-1}\Bigg(\sum\limits_{i=1}^{M_1} g(\hat{q}_i \cdot \widehat{\textrm{PIF}}_{1,i})\Bigg)
}
and a second \code{pif_global_ensemble}:
\deqn{
\widehat{\textrm{PIF}}_{2} = h^{-1}\Bigg(\sum\limits_{j=1}^{M_2} g(\hat{w}_j \cdot
\widehat{\textrm{PIF}}_{2,j})\Bigg)
}
This function computes the covariance between the first impact fraction and the
coefficients of the second fraction by computing:
\deqn{
\operatorname{Cov}(\widehat{\textrm{PIF}}_A, \widehat{\textrm{PIF}}_{B,j})
\approx \frac{1}{g'\big(\widehat{\textrm{PIF}}_A\big)}\sum\limits_{i=1}^{M_1}g'(\hat{q}_i
\cdot \widehat{\textrm{PIF}}_{A,i}) \Bigg[ \mathbb{E}[\hat{q}_i] \operatorname{Cov}\Big( \textrm{PIF}_{A,i},
\widehat{\textrm{PIF}}_{B,j}\Big) +  \mathbb{E}[\textrm{PIF}_{A,i}] \operatorname{Cov}\Big( \hat{q}_i,
\widehat{\textrm{PIF}}_{B,j}\Big) \Bigg]
}
where \eqn{\widehat{\textrm{PIF}}_{1,:} = (\widehat{\textrm{PIF}}_{1,1},
\widehat{\textrm{PIF}}_{1,2}, \dots, \widehat{\textrm{PIF}}_{1,M_1})^{\top}}
}

\seealso{
\code{\link[=cov_ensemble_weight]{cov_ensemble_weight()}}, \code{\link[=cov_atomic_pif]{cov_atomic_pif()}}
}
\keyword{internal}
