% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{cov_ensemble_weights}
\alias{cov_ensemble_weights}
\title{Covariance between the j-th weight of a \code{pif_global_ensemble_class} and
another \code{pif_global_ensemble_class}}
\usage{
cov_ensemble_weights(
  pif1,
  pif2,
  var_weights,
  var_pif_weights,
  recursive = !is.null(var_pif_weights)
)
}
\arguments{
\item{pif1}{A \code{pif_global_ensemble_class} from which the weight is taken.}

\item{pif2}{A \code{pif_global_ensemble_class} or \code{pif_atomic_class} to compute the covariance}

\item{var_weights}{Covariance matrix between the weights of
\code{pif1} and the weights of \code{pif2}. Entry \code{var_weights[i,j]} is the
covariance between the \code{i}th weight of \code{pif1} and the \code{j}th weight of \code{pif2}.}

\item{var_pif_weights}{Covariance structure between the potential
impact fractions in \code{pif1} and the weights in \code{pif2}.}
}
\description{
Calculates the covariance of the jth-weight
of a potential impact fraction of \code{pif_global_ensemble_class}
with a second \code{pif_global_ensemble_class} or \code{pif_atomic_class}.
}
\note{
The model currently works under the assumption that all \code{pif_atomic_class}
are independent from weights. Hence will return \code{0} if \code{pif2} is an atomic pif.
}
\section{Formula}{

Given a \code{pif_global_ensemble}:
\deqn{
\widehat{\textrm{PIF}}_{1} =
g^{-1}\Bigg(\sum\limits_{i=1}^{M_1} g(\hat{q}_i \cdot
\widehat{\textrm{PIF}}_{1,i})\Bigg)
}
and a second \code{pif_global_ensemble}:
\deqn{
\widehat{\textrm{PIF}}_{2} = h^{-1}\Bigg(\sum\limits_{j=1}^{M_2} h(\hat{w}_j \cdot
\widehat{\textrm{PIF}}_{2,j})\Bigg)
}
This function computes the covariance between the first impact fraction and the
weights of the second fraction. This returns a vector where the \code{j}th entry
is given by:
\deqn{
\operatorname{Cov}(\widehat{\textrm{PIF}}_A, \hat{w}_j) \approx
\frac{1}{g'\big(\widehat{\textrm{PIF}}_A\big)}\sum\limits_{i=1}^{M_1}g'(\hat{q}_i
\cdot \widehat{\textrm{PIF}}_{A,i})
\Bigg[ \hat{q}_i \operatorname{Cov}\Big(
\widehat{\textrm{PIF}}_{A,i}, \hat{w}_j\Big) +
\widehat{\textrm{PIF}}_{A,i} \operatorname{Cov}\Big( \hat{q}_i,
\hat{w}_j\Big) \Bigg]
}
where \eqn{\widehat{\textrm{PIF}}_{1,:} = (\widehat{\textrm{PIF}}_{1,1},
\widehat{\textrm{PIF}}_{1,2}, \dots, \widehat{\textrm{PIF}}_{1,M_1})^{\top}}
}

\examples{
\dontrun{
p1 <- paf(c(0.1, 0.2), beta = c(0.01, 0.14), var_p = diag(c(0.01, 0.011)),
  var_beta = 0, label ="1")
p2 <- paf(c(0.2, 0.3), beta = c(0.01, 0.15), var_p = diag(c(0.06, 0.052)),
  var_beta = 0, label = "2")
var_weights <- matrix(c(0.1, 0.01, -0.03, 0.01, 0.4, 0.02, -0.03, 0.02, 0.01), ncol = 3)
pt1 <- paf_total(p1, p2, weights = c(0.1, 0.9), var_weights = var_weights[1:2,1:2],
  label ="total")

#This gives 0 as the assumption is that atomic pifs don't correlate with anything
cov_ensemble_weights(p1, p2, var_weights = NULL, var_pif_weights = NULL)

#This returns the correlation of pt1 with its weights
cov_ensemble_weights(pt1, pt1, var_weights = NULL, var_pif_weights = NULL)

#Changes if it changes to ensemble
p3 <- paf(c(0.1, 0.2), beta = c(0.01, 0.14), var_p = diag(c(0.01, 0.011)),
  var_beta = 0, label = "3")
p4 <- paf(c(0.2, 0.3), beta = c(0.01, 0.15), var_p = diag(c(0.06, 0.052)),
  var_beta = 0, label = "4")
p5 <- paf(c(0.2, 0.3), beta = c(0.01, 0.15), var_p = diag(c(0.06, 0.052)),
  var_beta = 0, label = "5")
pe1 <- paf_ensemble(p3, p4, p5, weights = c(0.1, 0.5, 0.4),
  var_weights = var_weights, label = "ensemble")

#This returns the correlation of pt1 with its weights
cov_ensemble_weights(pe1, pe1, var_weights = NULL, var_pif_weights = NULL)

#You can also specify var_weights as either a matrix
var_weights     <- matrix(c(0.1, 0.2, 0.1, 0.3, 0.4, -0.2), ncol = 2)
var_pif_weights <- matrix(c(0.01, 0.03, 0.021, -0.02, 0.004, -0.5), ncol = 2)
cov_ensemble_weights(pe1, pt1, var_weights = var_weights, var_pif_weights = var_pif_weights)

#or a covariance structure:

}


}
\seealso{
\code{\link[=cov_atomic_pif]{cov_atomic_pif()}}, \code{\link[=cov_ensemble_atomic]{cov_ensemble_atomic()}} \code{\link[=cov_total_pif]{cov_total_pif()}}
}
\keyword{internal}
