% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{cov_total_pif}
\alias{cov_total_pif}
\title{Covariance function for a \code{pif_global_ensemble_class}}
\usage{
cov_total_pif(
  pif1,
  pif2,
  var_p = NULL,
  var_beta = NULL,
  var_weights = NULL,
  var_pif_weights = NULL,
  var_pifs = NULL,
  warning = TRUE
)
}
\arguments{
\item{pif1}{Either a \code{pif_atomic_class} or a \code{pif_global_ensemble_class}}

\item{pif2}{Either a \code{pif_atomic_class} or a \code{pif_global_ensemble_class}}

\item{var_p}{covariance matrix for the prevalences in both \code{pif1} and \code{pif2}.
Default \code{NULL} (no correlation).}

\item{var_beta}{covariance matrix for the parameter \code{beta} in both \code{pif1}
and \code{pif2}. Default \code{NULL} (no correlation).}

\item{var_weights}{Covariance between the weights of all the fractions
contained in \code{pif1} and all the fractions contained in \code{pif2}.}

\item{var_pif_weights}{Covariance vector between the weights in  \code{pif_ensemble} and
the \code{pif_atomic}. This refers to
the term \eqn{\operatorname{Cov}\Big( \hat{q}_i,\widehat{\textrm{PIF}}_{B,j}\Big)}
in the equation below. If set to \code{NULL} its automatically calculated.}

\item{var_pifs}{Covariance vector between the potential
impact fractions in \code{pif_ensemble} and \code{pif_atomic}. This refers to
the term \eqn{\operatorname{Cov}\Big( \textrm{PIF}_{A,i}, \widehat{\textrm{PIF}}_{B,j}\Big)}
in the equation below. If set to \code{NULL} its automatically calculated.}

\item{warning}{Boolean indicating whether to throw a warning if the labels on the
fractions involved are not unique.}
}
\description{
Recursively obtains the covariance matrix of a \code{pif_global_ensemble_class}
}
\section{Computation}{

This computes:
\deqn{
\operatorname{Cov}(\widehat{\textrm{PIF}}_{A}, \widehat{\textrm{PIF}}_{B})  \approx
\frac{1}{g'\big(\widehat{\textrm{PIF}}_A\big)}\sum\limits_{i=1}^{M_1}g'(\hat{q}_i \cdot
\widehat{\textrm{PIF}}_{A,i}) \Bigg[ \mathbb{E}[\hat{q}_i] \operatorname{Cov}\Big( \textrm{PIF}_{A,i},
\widehat{\textrm{PIF}}_{B}\Big) +  \mathbb{E}[\textrm{PIF}_{A,i}] \operatorname{Cov}\Big( \hat{q}_i,
\widehat{\textrm{PIF}}_{B}\Big) \Bigg]
}
where \eqn{\operatorname{Cov}\Big( \textrm{PIF}_{A,i},
\widehat{\textrm{PIF}}_{B}\Big)} is computed as in \code{\link[=cov_ensemble_atomic]{cov_ensemble_atomic()}}
in the case \eqn{\widehat{\textrm{PIF}}_{A,i}} is atomic or with the same formula
in the case it is an ensemble. The expression \eqn{\operatorname{Cov}\Big( \hat{q}_i,
\widehat{\textrm{PIF}}_{B}\Big) } is computed as in  \code{\link[=cov_ensemble_weights]{cov_ensemble_weights()}}
for each weight \eqn{q_j}.
}

\seealso{
\code{\link[=from_parameters_covariance_p_component]{from_parameters_covariance_p_component()}}, \code{\link[=cov_atomic_pif]{cov_atomic_pif()}},
\code{\link[=cov_ensemble_atomic]{cov_ensemble_atomic()}}, \code{\link[=cov_ensemble_weights]{cov_ensemble_weights()}}
}
\keyword{internal}
