% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-covariance_structures.R
\name{covariance_structure_class}
\alias{covariance_structure_class}
\title{Covariance structure class}
\usage{
covariance_structure_class(cov_list = list())
}
\arguments{
\item{cov_list}{List such that \code{list[[i]][[j]]} represents the covariance
between elements of the i-th potential impact fraction
and the j-th potential impact fraction.}
}
\description{
The covariance structure class represents a collection of
matrices and 0's such that \code{cov[[i]][[j]]} contains the
covariance of elements of the i-th potential impact fraction
and the j-th potential impact fraction.
}
