% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-covariance_generics.R
\name{covcor}
\alias{covcor}
\alias{covariance}
\alias{variance}
\alias{standard_deviation}
\alias{correlation}
\title{Covariance matrix, correlation matrix, variance and standard deviation
for potential impact fractions}
\usage{
covariance(
  x,
  ...,
  var_p = NULL,
  var_beta = NULL,
  var_weights = NULL,
  var_pif_weights = NULL,
  var_pifs = NULL,
  warning = FALSE
)

variance(x, ...)

standard_deviation(x, ...)

correlation(x, ..., var_p = NULL, var_beta = NULL)
}
\arguments{
\item{x}{A potential impact fraction}

\item{...}{Multiple additional potential impact fraction objects
separated by commas.}

\item{var_p}{covariance matrix for the prevalences in both \code{pif1} and \code{pif2}.
Default \code{NULL} (no correlation).}

\item{var_beta}{covariance matrix for the parameter \code{beta} in both \code{pif1}
and \code{pif2}. Default \code{NULL} (no correlation).}

\item{var_weights}{Covariance between the weights of all the fractions
contained in \code{pif1} and all the fractions contained in \code{pif2}.}

\item{var_pif_weights}{Covariance vector between the weights in  \code{pif_ensemble} and
the \code{pif_atomic}. This refers to
the term \eqn{\operatorname{Cov}\Big( \hat{q}_i,\widehat{\textrm{PIF}}_{B,j}\Big)}
in the equation below. If set to \code{NULL} its automatically calculated.}

\item{var_pifs}{Covariance vector between the potential
impact fractions in \code{pif_ensemble} and \code{pif_atomic}. This refers to
the term \eqn{\operatorname{Cov}\Big( \textrm{PIF}_{A,i}, \widehat{\textrm{PIF}}_{B,j}\Big)}
in the equation below. If set to \code{NULL} its automatically calculated.}

\item{warning}{Boolean indicating whether to throw a warning if the labels on the
fractions involved are not unique.}
}
\description{
Computes the covariance (\code{covariance}) or correlation (\code{correlation}) for multiple
potential impact fractions and the variance \code{variance} and standard deviation
\code{standard_deviation}for a potential impact fractions.
}
\examples{
# Get the approximate link_variance of a pif object
my_pif <- pif(p = 0.5, p_cft = 0.25, beta = 1.3,
              var_p = 0.1, var_beta = 0.2)
variance(my_pif)

# This is the same as link_covariance with just 1 pIF
covariance(my_pif)

# Calculate the link_covariance between 3 fractions with shared relative risk
beta <- 0.3
var_beta <- 0.1
pif1 <- pif(0.5, 0.2, beta, var_p = 0.5 * (1 - 0.5) / 100, var_beta = var_beta)
pif2 <- pif(0.3, 0.1, beta, var_p = 0.3 * (1 - 0.3) / 100, var_beta = var_beta)
pif3 <- pif(0.7, 0.3, beta, var_p = 0.7 * (1 - 0.7) / 100, var_beta = var_beta)
covariance(pif1, pif2, pif3)

# The link_covariance between a pif and itself only has the link_variance as entries
covariance(pif1, pif1)

# Or if there is a link_covariance structure between different betas you can specify with
# var_beta in the link_covariance
betas <- c(1.3, 1.2, 1.27)

# link_covariance among all betas
var_beta <- matrix(c(
  1.0000000, -0.12123053, 0.35429369,
  -0.1212305, 1.00000000, -0.04266409,
  0.3542937, -0.04266409, 1.00000000
), byrow = TRUE, ncol = 3)
pif1 <- pif(0.5, 0.2, betas[1], var_p = 0.5 * (1 - 0.5) / 100, var_beta = var_beta[1, 1])
pif2 <- pif(0.3, 0.1, betas[2], var_p = 0.3 * (1 - 0.3) / 100, var_beta = var_beta[2, 2])
pif3 <- pif(0.3, 0.1, betas[3], var_p = 0.3 * (1 - 0.3) / 100, var_beta = var_beta[3, 3])
covariance(pif1, pif2, pif3, var_beta = var_beta)

# Compute the correlation
correlation(pif1, pif2, pif3, var_beta = var_beta)
}
