% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fraction_type}
\alias{fraction_type}
\title{Get the type of the fraction}
\usage{
fraction_type(x)
}
\arguments{
\item{x}{A \code{pif_class} object}
}
\value{
Either \code{PIF} or \code{PAF} depending on the object
}
\description{
Obtain whether a fraction is a potential impact fraction or a
population attributable fraction
}
