% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-getters.R
\name{getters}
\alias{getters}
\alias{get_alpha_confint}
\alias{get_rr}
\alias{get_mu_obs}
\alias{get_mu_cft}
\alias{get_pif}
\alias{get_link_vals}
\alias{get_link_deriv_vals}
\alias{get_rr_link_deriv_vals}
\alias{get_variance_atomic}
\alias{get_ci}
\alias{get_pifs}
\alias{get_covariance}
\title{Getters}
\usage{
get_alpha_confint(self)

get_rr(self)

get_mu_obs(self)

get_mu_cft(self)

get_pif(self)

get_link_vals(self)

get_link_deriv_vals(self)

get_rr_link_deriv_vals(self)

get_variance_atomic(self)

get_ci(self)

get_pifs(self)

get_covariance(self)
}
\arguments{
\item{self}{A \code{pif_class} or \code{pif_atomic_class} object created with S7.}
}
\description{
Collection of functions for the potential impact fraction class
so that they work as getters of the properties. Each
function is constructed as \code{get_property}
}
\keyword{internal}
