% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{length}
\alias{length}
\title{Length of a \code{covariance_structure}}
\arguments{
\item{x}{A \code{pif_atomic_class}}
}
\description{
Gets the length of a covariance structure

Gets the length of a \code{pif_ensemble_class}

Gets the length of a \code{pif_atomic_class}
}
