% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{link_deriv_vals}
\alias{link_deriv_vals}
\title{Get the values of the derivative of link}
\usage{
link_deriv_vals(x)
}
\arguments{
\item{x}{A \code{pif_class} object}
}
\value{
A number indicating the derivative of \code{link(pif)}
}
\description{
Obtain the values of the derivative of the link function of
a potential impact fraction (PIF) or a population attributable
fraction (PAF) at \code{pif}
}
\examples{
#A potential impact fraction
pif1 <- pif(p = 0.2, p_cft = 0.1, beta = 1.2, quiet = TRUE)
link_deriv_vals(pif1)

}
