% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{mu_cft_fun}
\alias{mu_cft_fun}
\title{Compute the mean under the counterfactual prevalence}
\usage{
mu_cft_fun(p_cft, rr)
}
\arguments{
\item{p_cft}{Counterfactual prevalence (proportion) of the exposed
individuals for each of the \code{N} exposure levels.}

\item{rr}{The relative risk for each of the exposure levels.}
}
\value{
The mean relative risk under the counterfactual prevalence
}
\description{
Compute the mean under the counterfactual prevalence
}
\keyword{internal}
