% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{pif_fun}
\alias{pif_fun}
\title{Compute the potential impact fraction}
\usage{
pif_fun(p, p_cft, rr)
}
\arguments{
\item{p}{Prevalence (proportion) of the exposed individuals for
each of the \code{N} exposure levels.}

\item{p_cft}{Counterfactual prevalence (proportion) of the exposed
individuals for each of the \code{N} exposure levels.}

\item{rr}{The relative risk for each of the exposure levels.}
}
\value{
A potential impact fraction (numeric)
}
\description{
Compute the potential impact fraction
}
\keyword{internal}
