% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-utils.R
\name{pif_fun2}
\alias{pif_fun2}
\title{Compute the potential impact fraction}
\usage{
pif_fun2(mu_obs, mu_cft)
}
\arguments{
\item{mu_obs}{The average value of the relative risk in the observed population.}

\item{mu_cft}{The average value of the counterfactual relative risk in the population.}
}
\value{
A potential impact fraction (numeric)
}
\description{
Compute the potential impact fraction
}
\keyword{internal}
