% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pif.R
\name{pifpaf}
\alias{pifpaf}
\alias{paf}
\alias{pif}
\title{Potential Impact fraction and Population Attributable Fraction}
\usage{
paf(
  p,
  beta,
  sigma_p = NULL,
  sigma_beta = NULL,
  rr_link = "identity",
  rr_link_deriv = NULL,
  link = "log-complement",
  link_inv = NULL,
  link_deriv = NULL,
  conf_level = 0.95,
  quiet = FALSE
)

pif(
  p,
  p_cft = rep(0, length(p)),
  beta,
  sigma_p = NULL,
  sigma_beta = NULL,
  rr_link = "identity",
  rr_link_deriv = NULL,
  link = "log-complement",
  link_inv = NULL,
  link_deriv = NULL,
  conf_level = 0.95,
  type = "PIF",
  quiet = FALSE
)
}
\arguments{
\item{p}{Prevalence (proportion) of the exposed individuals for
each of the \code{N} exposure levels.}

\item{beta}{Relative risk parameter for which standard deviation is
available (usually its either the relative risk directly or the log
of the relative risk as most RRs, ORs and HRs come from exponential
models).}

\item{sigma_p}{Estimate of the colink_variance matrix of \code{p} where the entry
\code{sigma_p[i,j]} represents the colink_variance between \code{p[i]} and \code{p[j]}.}

\item{sigma_beta}{Estimate of the colink_variance matrix of \code{beta} where the entry
\code{sigma_beta[i,j]} represents the colink_variance between \code{beta[i]} and \code{beta[j]}.}

\item{rr_link}{Link function such that the relative risk is given by
\code{rr_link(beta)}.}

\item{rr_link_deriv}{Derivative of the link function for the relative risk.
The function tries to build it automatically from \code{rr_link} using
\code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}

\item{link}{Link function such that the \code{pif} confidence intervals
stays within the expected bounds.}

\item{link_inv}{The inverse of \code{link}. For example if \code{link}
is \code{logit} this should be \code{inv_logit}.}

\item{link_deriv}{Derivative of the \code{link} function. The function tries
to build it automatically from \code{link} using \code{\link[Deriv:Deriv]{Deriv::Deriv()}}.}

\item{conf_level}{Confidence level for the confidence interval (default 0.95).}

\item{quiet}{Whether to show messages.}

\item{p_cft}{Counterfactual prevalence (proportion) of the exposed
individuals for each of the \code{N} exposure levels.}

\item{type}{Character either Potential Impact Fraction (\code{PIF}) or
Population Attributable Fraction (\code{PAF})}
}
\description{
Calculates the potential impact fraction \code{pif} or the population
attributable fraction \code{paf} for a categorical exposure considering
an observed prevalence of  \code{p} and a relative risk
(or relative risk parameter) of \code{beta}.
}
\note{
This function assumes \code{p} and \code{beta} have been pre-computed from
the data and the individual-level data are not accessible to the
researchers. If either the data for the individual-level prevalence of
exposure \code{p} or the data for the individual-level risk estimate \code{beta}
can be accessed by the researcher other methods (such as the \code{pifpaf}
package should be preferred).
}
\section{Formulas}{

This function computed the potential impact fraction and its confidence
intervals using Walter's formula:
\deqn{
 \dfrac{
   \sum\limits_{i=1}^N p_i \text{RR}_i - \sum\limits_{i=1}^N p_i^{\text{cft}} \text{RR}_i
  }{
   \sum\limits_{i=1}^N p_i \text{RR}_i
  }
}
in the case of \code{N} exposure categories which is equivalent to Levine's formula
when there is only \code{1} exposure category:
\deqn{
 \dfrac{
   p (\text{RR} - 1) - p^{\text{cft}} (\text{RR} - 1)
  }{
   1 + p (\text{RR} - 1)
  }
}
}

\section{Link functions for the PIF}{

}

\section{Link functions for beta}{

}

\section{Population Attributable Fraction}{

The population attributable fraction corresponds to the potential impact
fraction at the theoretical minimum risk level. It is assumed that the
theoretical minimum risk level is a relative risk of 1. If no
counterfactual prevalence \code{p_cft} is specified, the model computes
the population attributable fraction.
}

\examples{
# This example comes from Levin 1953
# Relative risk of lung cancer given smoking was 3.6
# Proportion of individuals smoking where 49.9\%
# Calculates PAF (i.e. counterfactual is no smoking)
paf(p = 0.499, beta = 3.6)

# Assuming that beta and p had a link_variance
paf(p = 0.499, beta = 3.6, sigma_p = 0.001, sigma_beta = 1)

# If the link_variance was to high a logistic transform would be required
# Generates incorrect values for the interval:
paf(p = 0.499, beta = 3.6, sigma_p = 0.1, sigma_beta = 3)

# Logit fixes it
paf(p = 0.499, beta = 3.6, sigma_p = 0.1, sigma_beta = 3,
    link = "logit", quiet = TRUE)

# If the counterfactual was reducing the smoking population by 1/2
pif(p = 0.499, beta = 1.6, p_cft = 0.499/2, sigma_p = 0.001,
    sigma_beta = 1, link = "logit", quiet = TRUE)

}
