% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{print}
\alias{print}
\title{Print or show a potential impact fraction}
\arguments{
\item{x}{A \code{pif_class}}

\item{...}{Additional arguments to pass to \code{print} or \code{show}.}

\item{accuracy}{The accuracy of the printed value}
}
\description{
Function to print or show a potential impact fraction object
}
\examples{
my_pif <- pif(p = 0.2, beta = 1.3, var_beta = 0.1)
print(my_pif)

# Change the ammount of digits to show just 1
print(my_pif, accuracy = 0.1)
}
