% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{rowcol}
\alias{rowcol}
\alias{row_names}
\alias{col_names}
\title{Row and Column names}
\usage{
row_names(x)

col_names(x)
}
\arguments{
\item{x}{A \code{covariance_structure}}
}
\description{
Retrieve the row and column names of a \code{covariance_structure}
}
