% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-generics.R
\name{summary}
\alias{summary}
\title{Summary of a pif object}
\arguments{
\item{object}{A \code{pif_class} object.}

\item{...}{Additional parameters to pass to \code{summary} (ignored)}
}
\description{
Gets the potential impact fraction summary
}
\examples{
my_pif <- pif(p = 0.5, p_cft = 0.25, beta = 1.3, var_p = 0.1, var_beta = 0.2)
summary(my_pif)
}
