% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_product.R
\name{var_prod}
\alias{var_prod}
\title{Variance product formula}
\usage{
var_prod(x, y, var_x, var_y)
}
\description{
Implements the variance product formula. This is the formula for
\code{var(X*Y)} given by:
\deqn{
\text{Var}[XY] = Y^2\cdot\text{Var}[X]  + X^2\cdot \text{Var}[Y]
+ \text{Var}[X] \cdot \text{Var}[Y]
}
}
\keyword{internal}
